Module telamon::codegen

Helpers to generate code from an IR instance and fully specified decisions.

Structs

Dimension
An iteration dimension composed of one or mure fused dimensions.

Function
A function ready to execute on a device, derived from a constrained IR instance.

InductionLevel
An induction level associated to a dimension.

InductionVar
An induction variable, composed of multiple induction variable levels.

Instruction
An instruction to execute.

InternalMemBlock
A memory block allocated by the kernel.

NameMap
Maps variables to names.

Size
The size of an iteration dimension. The size is of the form: (factor * dividend_0 * dividend_1 * ...)) / divisor where the reminder of the division is null.

Enums

AllocationScheme
Indicates how is a memory block allocated.

Cfg
Represents a CFG of the targeted device.

MulMode

ParamVal
Represents the value of a parameter passed to the kernel by the host.

ParamValKey
Uniquely identifies a ParamVal.

Value
A value that can be named.

Traits

Namer
Assign names to variables.

Printer