Enum telamon::codegen::ParamVal

pub enum ParamVal<'a> {
    External(&'a Parameter, Type),
    Size(Size<'a>),
    GlobalMem(MemId, Size<'a>, Type),
}

Represents the value of a parameter passed to the kernel by the host.

Variants
External(&'a Parameter, Type)

A parameter given by the caller.

Size(Size<'a>)

A tiled dimension size computed on the host.

GlobalMem(MemId, Size<'a>, Type)

A pointer to a global memory block, allocated by the wrapper.

Methods

impl<'a> ParamVal<'a>

pub fn from_operand(operand: &'a Operand, space: &SearchSpace) -> Option<Self>

Builds the ParamVal needed to implement an operand, if any.

pub fn from_size(size: &Size<'a>) -> Option<Self>

Builds the ParamVal needed to get a size value, if any.

pub fn t(&self) -> Type

Returns the type of the parameter.

pub fn is_pointer(&self) -> bool

Indicates if the parameter is a pointer.

pub fn key(&self) -> ParamValKey

Returns a unique identifier for the ParamVal.