Struct telamon::codegen::Dimension

pub struct Dimension<'a> { /* fields omitted */ }

An iteration dimension composed of one or mure fused dimensions.

Methods

impl<'a> Dimension<'a>

pub fn id(&self) -> DimId

Returns the ID of the representant.

pub fn kind(&self) -> DimKind

Returns the kind of the dimension.

pub fn size(&self) -> &Size<'a>

Returns the size of the dimensions.

pub fn dim_ids(
&self
) -> impl Iterator<Item = DimId>

Returns the ids of the ir::Dimensions represented by this dimension.

pub fn induction_levels(&self) -> &[InductionLevel<'a>]

Returns the induction levels handled by this loop.

pub fn drain_induction_levels(&mut self) -> Vec<InductionLevel<'a>>

Gives the ownership on the induction levels computed by the dimension.

pub fn merge_from(&mut self, other: Self)

Merge another Dimension into this one.

pub fn host_values<'b>(
&'b self,
space: &'b SearchSpace
) -> impl Iterator<Item = ParamVal<'a>> + 'b

Returns the values to pass from the host to the device to implement self.