Struct telamon::codegen::Instruction

pub struct Instruction<'a> { /* fields omitted */ }

An instruction to execute.

Methods

impl<'a> Instruction<'a>

pub fn new(instruction: &'a Instruction<'a>, space: &SearchSpace) -> Self

Creates a new Instruction.

pub fn id(&self) -> InstId

Returns the ID of the instruction.

pub fn host_values(
&self,
space: &'a SearchSpace<'a>
) -> impl Iterator<Item = ParamVal<'a>>

Returns the values to pass from the host to implement this instruction.

pub fn t(&self) -> Option<Type>

Returns the type of the instruction.

pub fn operator(&self) -> &Operator

Returns the operator computed by the instruction.

pub fn instantiation_dims(&self) -> &[(DimId, u32)]

Returns the dimensions on which to instantiate the instruction.

pub fn as_reduction(&self) -> Option<(InstId, &DimMap)>

Indicates if the instruction performs a reduction, in wich case it returns the instruction that initializes the reduction, the DimMap to readh it and the reduction dimensions.

pub fn mem_flag(&self) -> Option<InstFlag>

Returns the memory flag of the intruction, if any.

pub fn has_side_effects(&self) -> bool

Indicates if the instruction has observable side effects.