Struct telamon::codegen::MemoryRegion

pub struct MemoryRegion<'a> { /* fields omitted */ }

A memory block allocated by the kernel.

Methods

impl<'a> MemoryRegion<'a>

pub fn new(
block: &'a Block,
num_threads_groups: &Option<Size<'a>>,
space: &'a SearchSpace
) -> Self

Creates a new MemoryRegion from an ir::Mem.

pub fn host_values(
&self,
space: &SearchSpace,
block_dims: &[Dimension<'a>]
) -> Vec<ParamVal<'a>>

Returns the value to pass from the host to the device to implement self.

pub fn id(&self) -> MemId

Returns the memory ID.

pub fn alloc_scheme(&self) -> AllocationScheme

Indicates how is the memory block allocated.

pub fn alloc_size(&self) -> Size<'a>

Generates the size of the memory to allocate.

pub fn local_size(&self) -> &Size<'a>

Returns the size of the part of the allocated memory accessible by each thread.

pub fn mem_space(&self) -> MemSpace

Returns the memory space the block is allocated in.

pub fn ptr_type(&self) -> Type

Returns the type of the pointer to the memory block.