Struct telamon::codegen::Size

pub struct Size<'a> { /* fields omitted */ }

The size of an iteration dimension. The size is of the form: (factor * dividend_0 * dividend_1 * ...)) / divisor where the reminder of the division is null.

Methods

impl<'a> Size<'a>

pub fn new(factor: u32, dividend: Vec<&'a Parameter>, divisor: u32) -> Self

Creates a new 'Size'.

pub fn from_ir(size: &PartialSize<'a>, _: &SearchSpace) -> Self

Converts an ir::Size to Self.

pub fn as_int(&self) -> Option<u32>

Returns the size of a dimension if it is staticaly known.

pub fn dividend(&self) -> &[&'a Parameter]

Returns the dividends.

pub fn divisor(&self) -> u32

Returns the divisor.

pub fn factor(&self) -> u32

Returns the factor.