Trait telamon::codegen::Printer

pub trait Printer<N: Namer> {
    fn get_int(n: u32) -> String;

    fn print_binop(
        &mut self, 
        vector_factors: [u32; 2], 
        op: BinOp, 
        operands_type: Type, 
        rounding: Rounding, 
        return_id: &str, 
        lhs: &str, 
        rhs: &str
    );

    fn print_unary_op(
        &mut self, 
        vector_factors: [u32; 2], 
        operator: UnaryOp, 
        operand_type: Type, 
        return_id: &str, 
        operand: &str
    );

    fn print_mul(
        &mut self, 
        vector_factors: [u32; 2], 
        return_type: Type, 
        round: Rounding, 
        mul_mode: MulMode, 
        return_id: &str, 
        op1: &str, 
        op2: &str
    );

    fn print_mad(
        &mut self, 
        vector_factors: [u32; 2], 
        ret_type: Type, 
        round: Rounding, 
        mul_mode: MulMode, 
        return_id: &str, 
        mlhs: &str, 
        mrhs: &str, 
        arhs: &str
    );

    fn print_ld(
        &mut self, 
        vector_factors: [u32; 2], 
        return_type: Type, 
        mem_space: MemSpace, 
        flag: InstFlag, 
        return_id: &str, 
        addr: &str
    );

    fn print_st(
        &mut self, 
        vector_factors: [u32; 2], 
        val_type: Type, 
        mem_space: MemSpace, 
        mem_flag: InstFlag, 
        predicate: Option<&str>, 
        addr: &str, 
        val: &str
    );

    fn print_label(&mut self, label_id: &str);

    fn print_cond_jump(&mut self, label_id: &str, cond: &str);

    fn print_sync(&mut self);

    fn name_operand<'a>(
        vector_levels: &[Vec<Dimension>; 2], 
        op: &Operand, 
        namer: &'a NameMap<N>
    ) -> Cow<'a, str>;

    fn name_inst<'a>(
        vector_levels: &[Vec<Dimension>; 2], 
        inst: InstId, 
        namer: &'a NameMap<N>
    ) -> Cow<'a, str>;

    fn print_lt_int(&mut self, t: Type, result: &str, lhs: &str, rhs: &str) { ... }

    fn print_equals(&mut self, t: Type, result: &str, lhs: &str, rhs: &str) { ... }

    fn print_add_int(&mut self, t: Type, result: &str, lhs: &str, rhs: &str) { ... }

    fn print_and(&mut self, t: Type, result: &str, lhs: &str, rhs: &str) { ... }

    fn print_move(&mut self, t: Type, result: &str, operand: &str) { ... }

    fn cfg_vec(&mut self, fun: &Function, cfgs: &[Cfg], namer: &mut NameMap<N>) { ... }

    fn cfg<'a>(&mut self, fun: &Function, c: &Cfg<'a>, namer: &mut NameMap<N>) { ... }

    fn parallel_induction_level(
        &mut self, 
        level: &InductionLevel, 
        namer: &NameMap<N>
    ) { ... }

    fn disable_threads<'a, I>(&mut self, threads: I, namer: &mut NameMap<N>)
    where
        I: Iterator<Item = &'a Dimension<'a>>,
    { ... }

    fn gen_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap<N>
    ) { ... }

    fn standard_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap<N>
    ) { ... }

    fn unroll_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap<N>
    ) { ... }

    fn privatise_global_block(
        &mut self, 
        block: &MemoryRegion, 
        namer: &mut NameMap<N>, 
        fun: &Function
    ) { ... }

    fn inst(
        &mut self, 
        vector_levels: &[Vec<Dimension>; 2], 
        inst: &Instruction, 
        namer: &mut NameMap<N>, 
        fun: &Function
    ) { ... }

    fn lower_type(t: Type, fun: &Function) -> Type { ... }

    fn mul_mode(from: Type, to: Type) -> MulMode { ... }
}

Required methods

fn get_int(n: u32) -> String

Get the representation of an integer in the target language.

fn print_binop(
&mut self,
vector_factors: [u32; 2],
op: BinOp,
operands_type: Type,
rounding: Rounding,
return_id: &str,
lhs: &str,
rhs: &str
)

Print return_id = lhs op rhs

fn print_unary_op(
&mut self,
vector_factors: [u32; 2],
operator: UnaryOp,
operand_type: Type,
return_id: &str,
operand: &str
)

Print return_id = op

fn print_mul(
&mut self,
vector_factors: [u32; 2],
return_type: Type,
round: Rounding,
mul_mode: MulMode,
return_id: &str,
op1: &str,
op2: &str
)

Print return_id = op1 * op2

fn print_mad(
&mut self,
vector_factors: [u32; 2],
ret_type: Type,
round: Rounding,
mul_mode: MulMode,
return_id: &str,
mlhs: &str,
mrhs: &str,
arhs: &str
)

Print return_id = mlhs * mrhs + arhs

fn print_ld(
&mut self,
vector_factors: [u32; 2],
return_type: Type,
mem_space: MemSpace,
flag: InstFlag,
return_id: &str,
addr: &str
)

Print return_id = load [addr]

fn print_st(
&mut self,
vector_factors: [u32; 2],
val_type: Type,
mem_space: MemSpace,
mem_flag: InstFlag,
predicate: Option<&str>,
addr: &str,
val: &str
)

Print store val [addr]

fn print_label(&mut self, label_id: &str)

print a label where to jump

fn print_cond_jump(&mut self, label_id: &str, cond: &str)

Print if (cond) jump label(label_id)

fn print_sync(&mut self)

Print wait on all threads

fn name_operand<'a>(
vector_levels: &[Vec<Dimension>; 2],
op: &Operand,
namer: &'a NameMap<N>
) -> Cow<'a, str>

Name an operand, vectorized on the given dimensions.

fn name_inst<'a>(
vector_levels: &[Vec<Dimension>; 2],
inst: InstId,
namer: &'a NameMap<N>
) -> Cow<'a, str>

Names an instruction, vectorized on the given dimensions.

Loading content...
Provided methods

fn print_lt_int(&mut self, t: Type, result: &str, lhs: &str, rhs: &str)

Prints a scalar less-than on integers.

fn print_equals(&mut self, t: Type, result: &str, lhs: &str, rhs: &str)

Prints a scalar equals instruction.

fn print_add_int(&mut self, t: Type, result: &str, lhs: &str, rhs: &str)

Prints a scalar addition on integers.

fn print_and(&mut self, t: Type, result: &str, lhs: &str, rhs: &str)

Prints an AND operation.

fn print_move(&mut self, t: Type, result: &str, operand: &str)

Prints a move instruction.

fn cfg_vec(&mut self, fun: &Function, cfgs: &[Cfg], namer: &mut NameMap<N>)

fn cfg<'a>(&mut self, fun: &Function, c: &Cfg<'a>, namer: &mut NameMap<N>)

Prints a cfg.

fn parallel_induction_level(
&mut self,
level: &InductionLevel,
namer: &NameMap<N>
)

Prints a multiplicative induction var level.

fn disable_threads<'a, I>(&mut self, threads: I, namer: &mut NameMap<N>) where
I: Iterator<Item = &'a Dimension<'a>>,

Change the side-effect guards so that the specified threads are disabled.

fn gen_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap<N>
)

Prints a Loop

fn standard_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap<N>
)

Prints a classic loop - that is, a sequential loop with an index and a jump to the beginning at the end of the block

fn unroll_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap<N>
)

Prints an unroll loop - loop without jumps

fn privatise_global_block(
&mut self,
block: &MemoryRegion,
namer: &mut NameMap<N>,
fun: &Function
)

fn inst(
&mut self,
vector_levels: &[Vec<Dimension>; 2],
inst: &Instruction,
namer: &mut NameMap<N>,
fun: &Function
)

Prints an instruction.

fn lower_type(t: Type, fun: &Function) -> Type

fn mul_mode(from: Type, to: Type) -> MulMode

Loading content...