Module telamon::device

Code generation and candidate evaluation for specific targets.

Modules

cuda
Defines the CUDA target.

x86
Defines the CPU target.

Enums

EvalMode
Indicates how evaluation should be performed.

Traits

ArgMap
Binds the argument names to their values.

ArrayArgument
Represents an array on the device.

AsyncEvaluator
An evaluation context that runs kernels asynchronously on the target device.

Context
Describes the context for which a function must be optimized.

Device
Holds the specifications of a target.

ScalarArgument
Represents a value that can be used as a Function argument. Must ensures the type is a scalar and does not contains any reference.

Functions

read_array
Copies the array to the host, interpreting it as an array of T.

write_array
Copies an values to the device array from the host array given as argument.

Type Definitions

AsyncCallback
A callback that is called after evaluating a kernel.