Struct telamon::device::cuda::Gpu

pub struct Gpu {
    pub name: String,
    pub sm_major: u8,
    pub sm_minor: u8,
    pub addr_size: u16,
    pub shared_mem_per_smx: u32,
    pub shared_mem_per_block: u32,
    pub allow_nc_load: bool,
    pub allow_l1_for_global_mem: bool,
    pub wrap_size: u32,
    pub thread_per_smx: u32,
    pub l1_cache_size: u32,
    pub l1_cache_line: u32,
    pub l2_cache_size: u32,
    pub l2_cache_line: u32,
    pub load_l2_latency: f64,
    pub load_ram_latency: f64,
    pub shared_bank_stride: u32,
    pub load_shared_latency: f64,
    pub num_smx: u32,
    pub max_block_per_smx: u32,
    pub smx_clock: f64,
    pub thread_rates: InstDesc,
    pub smx_rates: InstDesc,
    pub gpu_rates: InstDesc,
    pub add_f32_inst: InstDesc,
    pub add_f64_inst: InstDesc,
    pub add_i32_inst: InstDesc,
    pub add_i64_inst: InstDesc,
    pub mul_f32_inst: InstDesc,
    pub mul_f64_inst: InstDesc,
    pub mul_i32_inst: InstDesc,
    pub mul_i64_inst: InstDesc,
    pub mul_wide_inst: InstDesc,
    pub mad_f32_inst: InstDesc,
    pub mad_f64_inst: InstDesc,
    pub mad_i32_inst: InstDesc,
    pub mad_i64_inst: InstDesc,
    pub mad_wide_inst: InstDesc,
    pub div_f32_inst: InstDesc,
    pub div_f64_inst: InstDesc,
    pub div_i32_inst: InstDesc,
    pub div_i64_inst: InstDesc,
    pub syncthread_inst: InstDesc,
    pub loop_init_overhead: InstDesc,
    pub loop_iter_overhead: InstDesc,
    pub loop_end_latency: f64,
}

Represents CUDA GPUs.

Fields
name: String

The name of the GPU.

sm_major: u8

The compute capability major number.

sm_minor: u8

The compute capability minor number.

addr_size: u16

The size of pointers.

shared_mem_per_smx: u32

The amount of shared memory per SMX.

shared_mem_per_block: u32

The amount of shared memory available per block.

allow_nc_load: bool

true when non-coherent loads are enabled on the GPU.

allow_l1_for_global_mem: bool

ture when L1 caching is enabled for global memory accesses.

wrap_size: u32

The size of a wrap.

thread_per_smx: u32

The maximal number of resident thread per SMX.

l1_cache_size: u32

The size in bytes of the L1 cache.

l1_cache_line: u32

The size in bytes of a L1 cache line.

l2_cache_size: u32

The size in bytes of the L2 cache.

l2_cache_line: u32

The size in bytes of a L2 cache line.

load_l2_latency: f64

Latency of an L2 access.

load_ram_latency: f64

Latency of a RAM access.

shared_bank_stride: u32

The stride at wich replays occur in shared memory.

load_shared_latency: f64

Latency of a shared memory access.

num_smx: u32

The number of SMX in the GPU.

max_block_per_smx: u32

Maximum number of block per SMX.

smx_clock: f64

The clock of an SMX, in GHz.

thread_rates: InstDesc

Amount of processing power available on a single thread.

smx_rates: InstDesc

Amount of processing power available on a single SMX.

gpu_rates: InstDesc

Amount of processing power available on the whole GPU.

add_f32_inst: InstDesc add_f64_inst: InstDesc add_i32_inst: InstDesc add_i64_inst: InstDesc mul_f32_inst: InstDesc mul_f64_inst: InstDesc mul_i32_inst: InstDesc mul_i64_inst: InstDesc mul_wide_inst: InstDesc mad_f32_inst: InstDesc mad_f64_inst: InstDesc mad_i32_inst: InstDesc mad_i64_inst: InstDesc mad_wide_inst: InstDesc div_f32_inst: InstDesc div_f64_inst: InstDesc div_i32_inst: InstDesc div_i64_inst: InstDesc syncthread_inst: InstDesc loop_init_overhead: InstDesc

Overhead for entring the loop.

loop_iter_overhead: InstDesc

Overhead for a single iteration of the loop.

loop_end_latency: f64

Latency for exiting the loop.

Methods

impl Gpu

pub fn from_executor(executor: &Executor) -> Gpu

Returns the GPU model corresponding to `name.

pub fn dummy() -> Self

Creates a dummy GPU, to use for tests and benchmarks without evaluation.

pub fn print_ptx(&self, fun: &Function) -> String

Returns the PTX code for a Function.

pub fn blocks_per_smx(&self, space: &SearchSpace) -> u32

Computes the number of blocks that can fit in an smx.