Struct telamon::device::cuda::Kernel

pub struct Kernel<'a, 'b> { /* fields omitted */ }

An IR instance compiled into a CUDA kernel.

Methods

impl<'a, 'b> Kernel<'a, 'b>

pub fn compile(
fun: &'b Function<'b>,
gpu: &Gpu,
executor: &'a Executor,
opt_level: usize
) -> Self

Compiles a device function.

pub fn compile_remote(
function: &'b Function<'b>,
gpu: &Gpu,
executor: &'a Executor,
jit_daemon: &mut JITDaemon
) -> Self

Compiles a device function, using a separate process.

pub fn evaluate(&self, args: &Context) -> Result<u64, ()>

Runs a kernel and returns the number of cycles it takes to execute in cycles.

pub fn evaluate_real(&self, args: &Context, num_samples: usize) -> Vec<f64>

Runs a kernel and returns the number of cycles it takes to execute in nanoseconds, measured using cuda event rather than hardware counters.

pub fn gen_thunk(self, args: &'a Context) -> Thunk<'a>

Generates a Thunk than can then be run on the GPU.