Trait telamon::device::Argument

pub trait Argument: Sync + Send {
    fn t(&self) -> Type;

    fn raw_ptr(&self) -> *const c_void;

    fn size_of(&self) -> usize;

    fn as_size(&self) -> Option<u32> { ... }
}

Represents a value that can be used as a Function argument.

Required Methods

fn t(&self) -> Type

Returns the type of the argument.

fn raw_ptr(&self) -> *const c_void

Returns a raw pointer to the object.

fn size_of(&self) -> usize

Returns the size of the argument.

Provided Methods

fn as_size(&self) -> Option<u32>

Returns the argument interpreted as an iteration dimension size, if applicable.