Trait telamon::device::Context

pub trait Context: Sync {
    fn device(&self) -> &Device;

    fn evaluate(&self, space: &Function, mode: EvalMode) -> Result<f64, ()>;

    fn benchmark(&self, space: &Function, num_samples: usize) -> Vec<f64>;

    fn async_eval<'a, 'b>(
        &self, 
        num_workers: usize, 
        mode: EvalMode, 
        inner: &(Fn(&mut AsyncEvaluator<'a, 'b>) + Sync)
    );

    fn param_as_size(&self, name: &str) -> Option<u32>;

    fn eval_size(&self, size: &Size) -> u32 { ... }
}

Describes the context for which a function must be optimized.

Required Methods

fn device(&self) -> &Device

Returns the description of the device the code runs on.

fn evaluate(&self, space: &Function, mode: EvalMode) -> Result<f64, ()>

Returns the execution time of a fully specified implementation in nanoseconds.

This function should be called multiple times to obtain accurate execution time. Indeed, it only executes the code once, without warming the GPU first.

fn benchmark(&self, space: &Function, num_samples: usize) -> Vec<f64>

Compiles and benchmarks a functions. As opposed to Self::evaluate, the measured time contains potential startup times.

fn async_eval<'a, 'b>(
&self,
num_workers: usize,
mode: EvalMode,
inner: &(Fn(&mut AsyncEvaluator<'a, 'b>) + Sync)
)

Calls the inner closure in parallel, and gives it a pointer to an AsyncEvaluator to evaluate candidates in the context. skip_bad_bounds indicates than candidates whose bound is aboive the best candidate should be skiped.

fn param_as_size(&self, name: &str) -> Option<u32>

Returns a parameter interpreted as a size, if possible.

Provided Methods

fn eval_size(&self, size: &Size) -> u32

Evaluate a size.