Trait telamon::device::Device

pub trait Device: Sync {
    fn print(&self, function: &Function, out: &mut dyn Write);

    fn check_type(&self, t: Type) -> Result<(), TypeError>;

    fn max_block_dims(&self) -> u32;

    fn max_inner_block_size(&self) -> u32;

    fn max_threads(&self) -> u32;

    fn max_unrolling(&self) -> u32;

    fn has_vector_registers(&self) -> bool;

    fn can_vectorize(&self, dim: &Dimension, op: &Operator) -> bool;

    fn max_vectorization(&self, op: &Operator) -> [u32; 2];

    fn shared_mem(&self) -> u32;

    fn pointer_type(&self, mem_space: MemSpace) -> Type;

    fn supported_mem_flags(&self, op: &Operator) -> InstFlag;

    fn name(&self) -> &str;

    fn hw_pressure(
        &self, 
        space: &SearchSpace, 
        dim_sizes: &HashMap<DimId, Range>, 
        nesting: &HashMap<StmtId, Nesting>, 
        bb: &dyn Statement, 
        ctx: &dyn Context
    ) -> HwPressure;

    fn loop_iter_pressure(&self, kind: DimKind) -> (HwPressure, HwPressure);

    fn thread_rates(&self) -> HwPressure;

    fn block_rates(&self) -> HwPressure;

    fn total_rates(&self) -> HwPressure;

    fn bottlenecks(&self) -> &[&'static str];

    fn block_parallelism(&self, space: &SearchSpace) -> u32;

    fn additive_indvar_pressure(&self, t: &Type) -> HwPressure;

    fn multiplicative_indvar_pressure(&self, t: &Type) -> HwPressure;

    fn add_block_overhead(
        &self, 
        max_active_threads: FactorRange, 
        max_threads: FactorRange, 
        predication_factor: Range, 
        pressure: &mut HwPressure
    );

    fn lower_type(&self, t: Type, space: &SearchSpace) -> Option<Type>;

    fn gen_code(&self, implementation: &SearchSpace, out: &mut dyn Write) { ... }
}

Holds the specifications of a target.

Required methods

fn print(&self, function: &Function, out: &mut dyn Write)

Prints the code corresponding to a device Function.

fn check_type(&self, t: Type) -> Result<(), TypeError>

Indicates if a Type can be implemented on the device.

fn max_block_dims(&self) -> u32

Returns the maximal number of block dimensions.

fn max_inner_block_size(&self) -> u32

The maximal size inner block dimensions can have.

fn max_threads(&self) -> u32

Returns the maximal number of threads.

fn max_unrolling(&self) -> u32

Returns the maximal unrolling factor.

fn has_vector_registers(&self) -> bool

Indicates if the device uses vector registers or has imlicit gathers and scatters in vector instructions.

fn can_vectorize(&self, dim: &Dimension, op: &Operator) -> bool

Indicates if the operator can be vectorized along the dimension.

fn max_vectorization(&self, op: &Operator) -> [u32; 2]

Indicates the maximal vectorization factor for the given operator.

fn shared_mem(&self) -> u32

Returns the amount of shared memory available for each thread block.

fn pointer_type(&self, mem_space: MemSpace) -> Type

Indicates the type of the pointer for the given memory space.

fn supported_mem_flags(&self, op: &Operator) -> InstFlag

Indicates the memory flags supported by the operator.

fn name(&self) -> &str

Returns the name of the device.

fn hw_pressure(
&self,
space: &SearchSpace,
dim_sizes: &HashMap<DimId, Range>,
nesting: &HashMap<StmtId, Nesting>,
bb: &dyn Statement,
ctx: &dyn Context
) -> HwPressure

Returns the pressure cause by a Statement. For a dimension, returns the pressure for the full loop execution.

fn loop_iter_pressure(&self, kind: DimKind) -> (HwPressure, HwPressure)

Returns the pressure produced by a single iteration of a loop and the latency overhead of iterations.

fn thread_rates(&self) -> HwPressure

Returns the processing rates of a single thread, in units/ns

fn block_rates(&self) -> HwPressure

Returns the processing rates of a single block, in units/ns.

fn total_rates(&self) -> HwPressure

Returns the processing rates of the whole accelerator un units/ns.

fn bottlenecks(&self) -> &[&'static str]

Returns the names of potential bottlenecks.

fn block_parallelism(&self, space: &SearchSpace) -> u32

Returns the number of blocks that can be executed in parallel on the device.

fn additive_indvar_pressure(&self, t: &Type) -> HwPressure

Returns the pressure caused by an additive induction variable level.

fn multiplicative_indvar_pressure(&self, t: &Type) -> HwPressure

Returns the pressure caused by a multiplicative induction variable level.

fn add_block_overhead(
&self,
max_active_threads: FactorRange,
max_threads: FactorRange,
predication_factor: Range,
pressure: &mut HwPressure
)

Adds the overhead (per instance) due to partial wraps and predicated dimensions to the pressure. If the instruction is not predicated, predicated_dims_size should be 1.

fn lower_type(&self, t: Type, space: &SearchSpace) -> Option<Type>

Lowers a type using the memory space information. Returns None if some information is not yet specified.

Loading content...
Provided methods

fn gen_code(&self, implementation: &SearchSpace, out: &mut dyn Write)

Builds and outputs a constrained IR instance.

Loading content...