Trait telamon::device::ScalarArgument

pub unsafe trait ScalarArgument: Display + Send + Sync + 'static {
    fn t() -> Type
    where
        Self: Sized;

    fn get_type(&self) -> Type;

    fn raw_ptr(&self) -> *const c_void;

    fn as_operand<L>(&self) -> Operand<'static, L>
    where
        Self: Sized;

    fn gen_random<R: Rng>(_: &mut R) -> Self
    where
        Self: Sized;

    fn as_size(&self) -> Option<u32> { ... }
}

Represents a value that can be used as a Function argument. Must ensures the type is a scalar and does not contains any reference. Also must ensure that no two implementers should have the same ::t() value, because that is used for downcasting.

Required Methods

fn t() -> Type where
Self: Sized,

Returns the type of the argument.

fn get_type(&self) -> Type

Returns the type of the argument. This is a version of ::t() that can be called on a trait object.

fn raw_ptr(&self) -> *const c_void

Returns a raw pointer to the object.

fn as_operand<L>(&self) -> Operand<'static, L> where
Self: Sized,

Returns an operand holding the argument value as a constant.

fn gen_random<R: Rng>(_: &mut R) -> Self where
Self: Sized,

Generates a random instance of the argument type.

Provided Methods

fn as_size(&self) -> Option<u32>

Returns the argument interpreted as an iteration dimension size, if applicable.