Module telamon::explorer::bandit_arm

Structs

Env
An environment in which candidates can be refined.

ImplInfo
Informations on a fully-specified implementation

Node
Holds the children of a SubTree::InternalNode.

Path
Path to follow to reach a leaf in the tree.

TAGPolicy

TAGStats
Holds the TAG statistics for a given edge.

Tree
A search tree to perform a multi-armed bandit search.

UCTPolicy
TODO(bclement): The UCT formula is wrong, because 1) we are optimising as a reward while we actually have a cost and 2) the scale is wrong (evaluations are in the e6+ range but we do as if they were in 0-1).

UCTStats

Enums

DeadEndSource

TreeEvent
The possible tree events. WARNING: Changing the enums will break any pre-existing eventlog files. Adding new cases at the end only is safe.

Traits

TreePolicy
Policy to use when descending in the tree.