Struct telamon::explorer::bandit_arm::UCTPolicy

pub struct UCTPolicy { /* fields omitted */ }

TODO(bclement): The UCT formula is wrong, because 1) we are optimising as a reward while we actually have a cost and 2) the scale is wrong (evaluations are in the e6+ range but we do as if they were in 0-1).