Trait telamon::explorer::bandit_arm::TreePolicy

pub trait TreePolicy: Sized {
    type EdgeStats: Default;
    fn pick_child(
        &self, 
        env: &Env, 
        node: &Node<Self::EdgeStats>
    ) -> Option<usize>;

    fn backpropagate(&self, node: &Node<Self::EdgeStats>, idx: usize, eval: f64);
}

Policy to use when descending in the tree.

Associated Types

type EdgeStats: Default

Statistics stored on the edges.

Loading content...
Required methods

fn pick_child(&self, env: &Env, node: &Node<Self::EdgeStats>) -> Option<usize>

Pick a child in the given environment. Returns an index for the child, or None if the node has no children.

fn backpropagate(&self, node: &Node<Self::EdgeStats>, idx: usize, eval: f64)

Record an evaluation across an edge. This indicates that an evaluation of eval was found in a path which contains edge node.children[idx].

Loading content...