Module telamon::explorer::config

Defines a structure to store the configuration of the exploration. The configuration is read from the Setting.toml file if it exists. Some parameters can be overridden from the command line.

Structs

BanditConfig
Configuration parameters specific to the multi-armed bandit algorithm.

ChoiceOrdering
A list of ChoiceGroup representing the order in which we want to determine choices

Config
Stores the configuration of the exploration.

ParseChoiceGroupError
An error which can be returned when parsing a group of choices.

TAGConfig
Configuration for the TAG algorithm

UCTConfig
Configuration for the UCT algorithm

Enums

ChoiceGroup
An enum listing the Group of choices we can make For example, we can make first all DimKind decisions, then all Order decisions, etc.

Formula

NewNodeOrder
Indicates how to choose between nodes of the search tree when no children have been evaluated.

Normalization

Reward

SearchAlgorithm
Exploration algorithm to use.

TreePolicy
Tree policy configuration

ValueReduction