Enum telamon::explorer::config::ChoiceGroup

pub enum ChoiceGroup {
    LowerLayout,
    Size,
    DimKind,
    DimMap,
    Order,
    MemSpace,
    InstFlag,
}

An enum listing the Group of choices we can make For example, we can make first all DimKind decisions, then all Order decisions, etc.

Variants
LowerLayoutSizeDimKindDimMapOrderMemSpaceInstFlag