Enum telamon::explorer::config::NewNodeOrder

pub enum NewNodeOrder {
    Api,
    Random,
    Bound,
    WeightedRandom,
}

Indicates how to choose between nodes of the search tree when no children have been evaluated.

Variants
Api

Consider the nodes in the order given by the search space API.

Random

Consider the nodes in a random order.

Bound

Consider the nodes with the lowest bound first.

WeightedRandom

Consider the nodes with a probability proportional to the distance between the cut and the bound.

Methods

impl NewNodeOrder

pub fn pick_candidate<'a>(
self,
new_nodes: &[Candidate<'a>],
cut: f64
) -> Option<usize>

Called in montecarlo_descend, dispatch the choice of the next candidate according to our configuration

pub fn pick_index<IT>(self, nodes: IT, cut: f64) -> Option<usize> where
IT: Iterator<Item = (usize, f64)> + Clone,

Returns the index of the next candidate to consider.

impl NewNodeOrder

pub fn into_selector<IT, T>(self, cut: f64, bounds: IT) -> Option<Selector<T>> where
IT: Iterator<Item = (T, f64)>,