Enum telamon::explorer::config::OldNodeOrder

pub enum OldNodeOrder {
    Bandit,
    Bound,
    WeightedRandom,
}

Indicates how to choose between nodes of the search tree with at least one descendent evaluated.

Variants
Bandit

Use the weights from the bandit algorithm.

Bound

Take the candidate with the best bound.

WeightedRandom

Consider the nodes with a probability proportional to the distance between the cut and the bound.