Struct telamon::explorer::config::Config

pub struct Config {
    pub output_dir: String,
    pub log_file: String,
    pub event_log: String,
    pub num_workers: usize,
    pub stop_bound: Option<f64>,
    pub check_all: bool,
    pub timeout: Option<u64>,
    pub max_evaluations: Option<usize>,
    pub distance_to_best: Option<f64>,
    pub algorithm: SearchAlgorithm,
}

Stores the configuration of the exploration.

Fields
output_dir: String

Path to the output directory to use. All other paths (e.g. log_file) are relative to this directory. Defaults to the current working directory.

log_file: String

Name of the file in wich to store the logs.

event_log: String

Name of the file in which to store the binary event log.

num_workers: usize

Number of exploration threads.

stop_bound: Option<f64>

Indicates the search must be stopped if a candidate with an execution time better than the bound (in ns) is found.

check_all: bool

If true, check all implementations found for correctness. Otherwise (the default) only the best candidates are checked.

timeout: Option<u64>

Indicates the search must be stopped after the given number of minutes.

max_evaluations: Option<usize>

Indicates the search must be stopped after the given number of candidates have been evaluated.

distance_to_best: Option<f64>

A percentage cut indicate that we only care to find a candidate that is in a certain range above the best Therefore, if cut_under is 20%, we can discard any candidate whose bound is above 80% of the current best.

algorithm: SearchAlgorithm

Exploration algorithm to use. Needs to be last for TOML serialization, because it is a table.

Methods

impl Config

pub fn from_settings_toml() -> Self

Create a new configuration from the hardcoded Settings.toml configuration file if it exists.

pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Self, ConfigError>

Extract the configuration from the given configuration file.

pub fn from_json(json: &str) -> Result<Self, ConfigError>

Parse the configuration from a JSON string. Primary user is the Python API (through the C API).

pub fn output_path<P: AsRef<Path>>(&self, path: P) -> Result<PathBuf>

pub fn create_log(&self) -> Result<BufWriter<File>>

pub fn create_eventlog(&self) -> Result<Writer<EventLog>>