Enum telamon::explorer::mcts::Message

pub enum Message {
    Node {
        id: NodeId,
        parent: Option<(NodeId, EdgeIndex)>,
        children: Vec<Action>,
        bound: Option<Bound>,
        discovery_time: Duration,
    },
    Trace {
        thread: String,
        events: Vec<Timed<Event>>,
    },
    Evaluation {
        id: NodeId,
        value: Option<f64>,
        result_time: Duration,
    },
}

A log message.

The way that the log is generated ensures that all nodes referenced in a Trace or Evaluation event will have had a corresponding Node event previously in the log.

Variants
Node

A new node was created

Fields of Node

id: NodeId

Id of the newly created node

parent: Option<(NodeId, EdgeIndex)>

Parent of the node, if it exists. This is a pair containing the node identifier of the parent and index in the children list.

children: Vec<Action>

All available children for this node. This includes all actions ever considered, even those that end up being forbidden due to constraint propagation.

bound: Option<Bound>

Explained bound from the performance model. None if the node was killed due to constraint propagation, in which case we can't run the performance model.

discovery_time: Duration

Time at which the node was discovered.

Trace

Sequence of actions (moves in the tree) performed by a specific thread. Starts at the root of the tree.

Fields of Trace

thread: String

The thread performing the descent. There can be multiple traces per thread, and they will share the thread field.

events: Vec<Timed<Event>>

Sequence of events performed by the thread.

Evaluation

A node was evaluated

Fields of Evaluation

id: NodeId

Identifier of the evaluated node

value: Option<f64>

Evaluation result. If None, the node was cut at evaluation time or otherwise timed out.

result_time: Duration

Time at which the evaluation results were made available and backpropagation started.