Struct telamon::helper::SignatureBuilder

pub struct SignatureBuilder<'a, AM> where
    AM: ArgMap + Context + 'a,  { /* fields omitted */ }

Helper struct to build a Signature.

Methods

impl<'a, AM> Builder<'a, AM> where
AM: ArgMap + Context + 'a,

pub fn new(name: &str, context: &'a mut AM) -> Self

Creates a new builder for a function with the given name.

pub fn scalar<T: ScalarArgument>(&mut self, name: &str, arg: T)

Creates a new parameter and binds it to the given value.

pub fn array<S: ScalarArgument>(
&mut self,
name: &str,
size: usize
) -> (Id, Arc<AM::Array>)

Creates a new parameter and binds it to a freshly allocated an array.

pub fn tensor<'b, S: ScalarArgument>(
&mut self,
name: &'b str,
dim_sizes: Vec<DimSize<'b>>,
read_only: bool
) -> Tensor<'b, S> where
<AM as ArgMap>::Array: 'b,

Allocates an n-dimensional array.

pub fn eval_size(&self, size: DimSize) -> u32

Evaluates a size in the context.

pub fn get(self) -> Signature

Returns the Signature created by the builder.