Struct telamon::helper::tensor::Tensor

pub struct Tensor<'a, S: ScalarArgument> { /* fields omitted */ }

A tensor allocated in main memory.

Methods

impl<'a, S> Tensor<'a, S> where
S: ScalarArgument,

pub fn new(
name: &'a str,
dim_sizes: Vec<DimSize<'a>>,
read_only: bool,
mem_id: MemId,
array: Arc<ArrayArgument + 'a>
) -> Self

Allocates a new Tensor in the context.

pub fn load(&self, tiling: &[&[u32]], builder: &mut Builder) -> VirtualTensor

Creates a VirtualTensor that contains the values of self, loaded in registers.

pub fn read_to_host(&self, context: &Context) -> ArrayD<S>

Reads the tensor value in the context and copies it on the host.