Struct telamon::helper::tensor::TensorBuilder

pub struct TensorBuilder<'a> { /* fields omitted */ }

An helper to build a tensor.

Methods

impl<'a> TensorBuilder<'a>

pub fn new(name: &'a str, storage_dims: Vec<DimSize<'a>>) -> Self

Start building a Tensor with the given logical layout.

pub fn transpose(&mut self, lhs: usize, rhs: usize) -> &mut Self

Swap two dimensions in the memory layout of the tensor. Keeps the logical layout untouched.

pub fn stride_dim(&mut self, dim: usize) -> &mut Self

Removes a logical dimension but keeps it in the storage.

pub fn enable_writes(&mut self) -> &mut Self

Allows writing to the tensor.

pub fn finish<S, AM>(&self, builder: &mut SignatureBuilder<AM>) -> Tensor<'a, S> where
S: ScalarArgument,
AM: ArgMap + Context + 'a,

Builds the Tensor.