Struct telamon::helper::tensor::VirtualTensor

pub struct VirtualTensor { /* fields omitted */ }

A tensor loaded in registers.

Methods

impl VirtualTensor

pub fn new(inst: InstId, dims: Vec<LogicalDim>) -> Self

Creates a new VirtualTensor.

pub fn dim_map(
&self,
dims: &[&LogicalDim],
scope: DimMapScope<()>,
builder: &mut Builder
) -> Operand<()>

Creates an operand that yeilds the values of the tensor in the given loop nest.

pub fn store<S>(
&self,
tensor: &Tensor<S>,
builder: &mut Builder
) -> VirtualTensor where
S: ScalarArgument,

Stores the VirtualTensor in memory. Stores contiguously without taking the layout of the target tensor into account.

pub fn inst(&self) -> InstId

Returns the underlying instruction.

pub fn num_dims(&self) -> usize

Returns the number of logical dimensions.

pub fn same_shape<T>(&self, other: &Self, function: &Function<T>) -> bool

Returns true if the other cirtual tensor has the same number of dimensions and each dimension has the same size

pub fn iter(&self) -> Iter<LogicalDim>