Enum telamon::ir::Operator

pub enum Operator<'a, L = LoweringMap> {
    BinOp(BinOp, Operand<'a, L>, Operand<'a, L>, Rounding),
    Mul(Operand<'a, L>, Operand<'a, L>, Rounding, Type),
    Mad(Operand<'a, L>, Operand<'a, L>, Operand<'a, L>, Rounding),
    Mov(Operand<'a, L>),
    Ld(Type, Operand<'a, L>, AccessPattern<'a>),
    St(Operand<'a, L>, Operand<'a, L>, bool, AccessPattern<'a>),
    TmpLd(Type, MemId),
    TmpSt(Operand<'a, L>, MemId),
    Cast(Operand<'a, L>, Type),
}

The operation performed by an instruction.

Variants
BinOp(BinOp, Operand<'a, L>, Operand<'a, L>, Rounding)

A binary arithmetic operator.

Mul(Operand<'a, L>, Operand<'a, L>, Rounding, Type)

Performs a multiplication with the given return type.

Mad(Operand<'a, L>, Operand<'a, L>, Operand<'a, L>, Rounding)

Performs s multiplication between the first two operands and adds the result to the third.

Mov(Operand<'a, L>)

Moves a value into a register.

Ld(Type, Operand<'a, L>, AccessPattern<'a>)

Loads a value of the given type from the given address.

St(Operand<'a, L>, Operand<'a, L>, bool, AccessPattern<'a>)

Stores the second operand at the address given by the first. The boolean specifies if the instruction has side effects. A store has no side effects when it writes into a cell that previously had an undefined value.

TmpLd(Type, MemId)

Represents a load from a temporary memory that is not fully defined yet.

TmpSt(Operand<'a, L>, MemId)

Represents a store to a temporary memory that is not fully defined yet.

Cast(Operand<'a, L>, Type)

Casts a value into another type.

Methods

impl<'a, L> Operator<'a, L>

pub fn check(
&self,
iter_dims: &HashSet<DimId>,
fun: &Function<L>
) -> Result<(), Error>

Ensures the types of the operands are valid.

pub fn t(&self) -> Option<Type>

Returns the type of the value produced.

pub fn operands(&self) -> Vec<&Operand<'a, L>>

Retruns the list of operands.

pub fn operands_mut<'b>(&'b mut self) -> Vec<&'b mut Operand<'a, L>>

Retruns the list of mutable references to operands.

pub fn has_side_effects(&self) -> bool

Returns true if the operator has side effects.

pub fn merge_dims(&mut self, lhs: DimId, rhs: DimId)

Renames a basic block.

pub fn mem_access_pattern(&self) -> Option<Cow<AccessPattern>>

Returns the pattern of access to the memory by the instruction, if any.

pub fn mem_used(&self) -> Option<MemId>

Returns the memory blocks referenced by the instruction.

pub fn supports_nc_access(&self) -> bool

Indicates if the operator supports non-coherent memory accesses.

pub fn map_operands<T, F>(self, f: F) -> Operator<'a, T> where
F: FnMut(Operand<'a, L>) -> Operand<'a, T>,

impl<'a> Operator<'a, ()>

pub fn freeze(self, cnt: &mut Counter) -> Operator<'a>