Enum telamon::ir::TypeError

pub enum TypeError {
    InvalidType {
        t: Type,
    },
    ExpectedReturnType {
        inst: InstId,
    },
    InvalidRounding {
        rounding: Rounding,
        t: Type,
    },
    WrongType {
        given: Type,
        expected: ExpectedType,
    },
    UnexpectedType {
        t: Type,
    },
}

Errors that can be raised when creating an IR instance.

Variants
InvalidType
Fields of InvalidType

t: Type
ExpectedReturnType
Fields of ExpectedReturnType

inst: InstId
InvalidRounding
Fields of InvalidRounding

rounding: Roundingt: Type
WrongType
Fields of WrongType

given: Typeexpected: ExpectedType
UnexpectedType
Fields of UnexpectedType

t: Type

Methods

impl TypeError

pub fn check_equals(given: Type, expected: Type) -> Result<(), Self>

Ensures a type is equal to the expected one.

pub fn check_integer(given: Type) -> Result<(), Self>

Ensures the given type is an integer type.

pub fn check_float(given: Type) -> Result<(), Self>

Ensures the given type is a floating point type.