Struct telamon::ir::mem::Block

pub struct Block { /* fields omitted */ }

A block of memory allocated on the device by the kernel.

Methods

impl Block

pub fn mem_id(&self) -> MemId

The ID of the block.

pub fn maps_dims(&self, lhs: DimId, rhs: DimId) -> bool

Indicates if two dimensions are mapped by a temporary memory block.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn mapped_dims(&self) -> &[(DimId, DimId)]

Returns the list of mapped dimensions.

pub fn is_private(&self) -> bool

Indicates if the block is privatised per block of thread.

pub fn base_size(&self) -> u32

Return the base size of the block, if it is statically known.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn uses(&self) -> &[InstId]

The list of instructions referencing the memory block.

pub fn add_use(&mut self, inst: InstId)

Add a use to the block.