Struct telamon::ir::Counter

pub struct Counter {
    pub next_mem: usize,
    pub next_inst: usize,
    pub next_dim: usize,
    pub next_dim_mapping: u16,
}

A wrapper used to count extra dimensions that will be needed in the future and allocates IDs for them. This is used when freezing in order to pre-allocate IDs for the various objects (internal memory block, instructions, dimensions, etc.) required for future lowering.

Fields
next_mem: usize next_inst: usize next_dim: usize next_dim_mapping: u16
Methods

impl Counter

pub fn next_mem(&mut self) -> InternalId

pub fn next_inst(&mut self) -> InstId

pub fn next_dim(&mut self) -> DimId

pub fn next_dim_mapping(&mut self) -> DimMappingId