Struct telamon::ir::Dimension

pub struct Dimension<'a, L = LoweringMap> { /* fields omitted */ }

Represents an iteration dimension.

Methods

impl<'a> Dimension<'a, ()>

pub fn freeze(self) -> Dimension<'a>

Sets the dimension as frozen.

impl<'a, L> Dimension<'a, L>

pub fn new(
id: DimId,
size: PartialSize<'a>,
logical_dim: Option<LogicalDimId>
) -> Result<Self, Error>

Creates a new dimension.

pub fn new_static(
id: DimId,
possible_sizes: VecSet<u32>,
logical_dim: Option<LogicalDimId>
) -> Result<Self, Error>

Creates a dimension with a statically known size, picked in a list of possibilities.

pub fn with_same_size(id: DimId, other: &Self) -> Self

Creates a new dimension with the same size as an existing one.

pub fn size(&self) -> &PartialSize<'a>

Retruns the size of the dimension.

pub fn possible_sizes(&self) -> Option<&[u32]>

Returns the values the size can take, if it is statically known.

pub fn id(&self) -> DimId

Returns the id of the dimension.

pub fn iterated(
&self
) -> impl Iterator<Item = InstId> + '_

Returns the constructs iterated along this dimension.

pub fn add_iterated(&mut self, inst: InstId)

Adds a stmt that is iterated along self.

pub fn is_thread_dim(&self) -> bool

Indicates if the dimension is a thread dimension.

pub fn set_thread_dim(&mut self)

Sets the dimension as a thread dimension.

pub fn logical_dim(&self) -> Option<LogicalDimId>

Returns the logical dimension this dimension is part of, if any.

pub fn dim_mappings(&self) -> &VecSet<DimMappingId>

Returns the list of dimensions mapping containing this one.

pub fn register_dim_mapping(&mut self, mapping: &DimMapping)

Register a dimension mapping.

pub fn inner_vars(&self) -> &VecSet<VarId>

Returns the list of variables available inside the dimension.

pub fn register_inner_var(&mut self, var: VarId)

Register a variable available inside the dimension.

pub fn set_sequential(&mut self)

Indicates the dimension cannot be parallelized.

pub fn is_parallelizable(&self) -> bool

Indicates if the dimension can be parallelized.