Struct telamon::ir::Instruction

pub struct Instruction<'a, L = LoweringMap> { /* fields omitted */ }

Represents an instruction.

Methods

impl<'a, L> Instruction<'a, L>

pub fn new(
operator: Operator<'a, L>,
id: InstId,
iter_dims: HashSet<DimId>,
fun: &Function<L>
) -> Result<Self, Error>

Creates a new instruction and type-check the operands.

pub fn operands(&self) -> Vec<&Operand<'a, L>>

Returns an iterator over the operands of this instruction.

pub fn t(&self) -> Option<Type>

Returns the type of the variable produced by an instruction.

pub fn operator(&self) -> &Operator<L>

Returns the operator of the instruction.

pub fn id(&self) -> InstId

Returns the InstId representing the instruction.

pub fn has_side_effects(&self) -> bool

Returns true if the instruction has side effects.

pub fn lower_layout(
&mut self,
ld_idx: Operand<'a, L>,
ld_pattern: AccessPattern<'a>,
st_idx: Operand<'a, L>,
st_pattern: AccessPattern<'a>
) where
L: Clone,

Applies the lowering of a layout to the instruction.

pub fn dim_maps_to_lower(&self, lhs: DimId, rhs: DimId) -> Vec<usize>

Indicates the operands for wich a DimMap must be lowered if lhs and rhs are not mapped.

pub fn as_mem_inst(&self) -> Option<&Instruction<L>>

Returns 'self' if it is a memory instruction.

pub fn as_reduction(&self) -> Option<(InstId, &DimMap, &[DimId])>

Indicates if the instruction performs a reduction.

pub fn is_reduction_common_dim(&self, init: InstId, dim: DimId) -> bool

Returns 'true' if self is a reduction initialized by init, and if 'dim' should have the same nesting with 'init' that with 'self'.

pub fn merge_dims(&mut self, lhs: DimId, rhs: DimId)

Rename a dimension to another ID.

pub fn iteration_dims(&self) -> &HashSet<DimId>

The list of dimensions the instruction must be nested in.

pub fn add_iteration_dimension(&mut self, dim: DimId) -> bool

Adds a new iteration dimension. Indicates if the dimension was not already an iteration dimension.

pub fn result_variable(&self) -> Option<VarId>

Returns the Variable holding the result of this instruction.

pub fn set_result_variable(&mut self, variable: VarId)

Sets the Variable holdings the result of this instruction.

impl<'a> Instruction<'a, ()>

pub fn freeze(self, cnt: &mut Counter) -> Instruction<'a>

impl<'a> Instruction<'a>

pub fn lower_dim_map(
&mut self,
op_id: usize,
new_src: InstId,
new_dim_map: DimMap
)

Lowers the DimMap of an operand into an access to a temporary memory.