Struct telamon::ir::LoweredDimMap

pub struct LoweredDimMap {
    pub mem: InternalId,
    pub store: InstId,
    pub load: InstId,
    pub st_dims_mapping: Vec<(DimMappingId, [DimId; 2])>,
    pub ld_dims_mapping: Vec<(DimMappingId, [DimId; 2])>,
}

A point-to-point communication lowered into a store and a load.

Fields
mem: InternalId store: InstId load: InstId st_dims_mapping: Vec<(DimMappingId, [DimId; 2])>

Mapping from production dimensions to store dimensions.

ld_dims_mapping: Vec<(DimMappingId, [DimId; 2])>

Mapping from consumption dimensions to load dimensions.

Methods

impl LoweredDimMap

pub fn register_new_objs(&self, fun: &Function, new_objs: &mut NewObjs)

Adds the objects created by the lowering to the list of new objects.

pub fn mem_dimensions(&self) ->

Returns the dimensions of the memory layout to create. For each dimension, gives a pair (store dim, load dim).

pub fn store_dims(&self) ->

Returns the dimensions that store the value.

pub fn load_dims(&self) ->

Returns the dimensions that load the value.