Struct telamon::model::Nesting

pub struct Nesting {
    pub inner_dims: VecSet<Id>,
    pub inner_bbs: VecSet<BBId>,
    pub outer_dims: VecSet<Id>,
    pub before_self: VecSet<Id>,
    pub after_self: VecSet<Id>,
    pub bigger_merged_dims: VecSet<Id>,
    pub num_unmapped_threads: u32,
    pub max_threads_per_block: u64,
    // some fields omitted
}

Nesting of an object.

Fields
inner_dims: VecSet<Id>

Dimensions nested inside the current BB.

inner_bbs: VecSet<BBId>

Basic blocks nested inside the current BB.

outer_dims: VecSet<Id>

Dimensions nested outsidethe current BB.

before_self: VecSet<Id>

Dimensions to be processed before the current BB.

after_self: VecSet<Id>

Dimensions that should not take the current BB into account when processed.

bigger_merged_dims: VecSet<Id>

The dimensions that can be merged to this one and have a bigger ID.

num_unmapped_threads: u32

Number of threads that are not represented in the active dimensions of the block.

max_threads_per_block: u64

Maximal number of threads this block can be in, considering only outer and mapped out dimensions.