Struct telamon::offline_analysis::tree::CandidateNode

pub struct CandidateNode { /* fields omitted */ }

A node of in the resconstructed candidate tree

Methods

impl CandidateNode

A node in the re-created tree representing a candidate

pub fn discovery_time(&self) -> Duration

Returns the time at which the node was discovered

pub fn internal_time(&self) -> Option<Duration>

Returns the time at which the node was marked as an internal node

pub fn rollout_time(&self) -> Option<Duration>

Returns the time at which the node was marked as a rollout node

pub fn implementation_time(&self) -> Option<Duration>

Returns the time at which the node was marked as an implementation

pub fn deadend_time(&self) -> Option<Duration>

Returns the time at which the node was marked as a deadend

pub fn bound(&self) -> Ref<Option<Bound>>

Returns the bound from the performance model for this candidate

pub fn score(&self) -> Option<f64>

Returns the score from the evaluation of this candidate

pub fn parent(&self) -> Option<CandidateNode>

Returns the parent node or None if this is the root node

pub fn is_explored(&self) -> bool

Returns true if this node has been explored (i.e., it is a rollout node or an internal node)

pub fn child(&self, child_idx: usize) -> Option<CandidateNode>

Returns the child at index child_idx. This may be None if the child hasn't been set before.

Panics

Panics if the index is invalid.

pub fn children(&self) -> impl Iterator<Item = Option<CandidateNode>>

Returns an interator that allows for iteration over all children of this node as Option, including unexplored children represented by None values.

pub fn num_children(&self) -> usize

Returns the number of children, including unexplored children

pub fn id(&self) -> NodeId

Returns the ID of this node

pub fn is_root(&self) -> bool

Indicates whether this is the virtual root node

pub fn action(&self) -> Option<Action>

Returns the action associated to the edge from the parent of this node to the node. This might be None if the node does not have a parent (i.e., if this is the root).

pub fn action_str(&self) -> String

Returns the action associated to the edge from the parent of this node to the node as a string. If no action is associated with the edge, an empty string is returned.

pub fn is_deadend(&self) -> bool

Returns true if this node was declared as a deadend by at least one thread

pub fn is_rollout_node(&self) -> bool

Returns true if this node was declared as a rollout node by at least one thread

pub fn is_internal_node(&self) -> bool

Returns true if this node was declared as an internal node by at least one thread

pub fn is_implementation(&self) -> bool

Returns true if this node was declared as an implementation by at least one thread

pub fn declare_deadend(&mut self, timestamp: Duration)

Marks this node as a deadend. The time passed as a parameter becomes the deadend time if no other thread has declared the node as a deadend with a lower timestamp.

pub fn declare_internal(&mut self, timestamp: Duration)

Marks this node as an internal node. The time passed as a parameter becomes the internal time if no other thread has declared the node as internal with a lower timestamp.

pub fn declare_rollout(&mut self, timestamp: Duration)

Marks this node as a rollout node. The time passed as a parameter becomes the rollout time if no other thread has declared the node as a rollout node with a lower timestamp.

pub fn declare_implementation(&mut self, timestamp: Duration)

Marks this node as an implementation. The time passed as a parameter becomes the implementation time if no other thread has declared the node as an implementation with a lower timestamp.

pub fn set_score(&mut self, score: f64)

Sets the score from an evaluation

Panics

Panics if the score has been set beforehand