Struct telamon::offline_analysis::tree::CandidateTree

pub struct CandidateTree { /* fields omitted */ }

A reconstructed tree

Methods

impl CandidateTree

pub fn new() -> Self

Creates a new, empty tree

pub fn get_root(&self) -> CandidateNode

Returns the root node

Panics

Panics if no root has been created beforehand

pub fn get_node(&self, id: NodeId) -> CandidateNode

Returns the node with the given id

Panics

Panics if no such node exists in the tree.

pub fn has_node(&self, id: NodeId) -> bool

Checks whether a node with the given id exists in the tree.

pub fn extend(
&mut self,
node_id: NodeId,
discovery_time: Duration,
parent: Option<(NodeId, EdgeIndex)>,
bound: Option<Bound>,
actions: &mut Vec<Action>
)

Creates a new candidate node with the ID node_id and inserts it into the tree. If the parent is not None, a parent-child relationship between the parent and the new node is set up. The parameter bound is the optional bound for this candidate from the performance model. Actions is an array of actions associated to the outgoing edges of this candidate to its children and may be empty if this node does not have any children in the final recreated tree.

Automatically sets the root of the tree to the newly created node if parent is None.

Panics

Panics If parent is not None and the ID provided for the parent node is unknown.