Enum telamon::search_space::Action

#[repr(C)]

pub enum Action {
    InstFlag(InstId, InstFlag),
    Size(DimId, NumericSet),
    IsThreadDim(DimId, Bool),
    ThreadMapping(DimId, DimId, ThreadMapping),
    IsIterationDim(InstId, DimId, Bool),
    MemorySpace(VarId, MemorySpace),
    MemSpace(MemId, MemSpace),
    DimKind(DimId, DimKind),
    Order(StmtId, StmtId, Order),
    DimMapping(DimId, DimId, DimMapping),
    IncrementTilingFactor(LogicalDimId, DimId, Bool),
    TilingFactor(LogicalDimId, Range),
    IsThreadDimClassCounter(DimId, Range),
    NumThreads(HalfRange),
    NumThreadDims(HalfRange),
    IsIterationDimClassCounter(InstId, DimId, Range),
    IncrementMemSize(MemId, DimId, DimId, Bool),
    MemSize(MemId, HalfRange),
    SharedMemUsed(HalfRange),
    IncrementUnrollFactor(InstId, DimId, Bool),
    UnrollFactor(InstId, HalfRange),
    IncrementInnerVectorFactor(InstId, DimId, Bool),
    InnerVectorFactor(InstId, HalfRange),
    IncrementOuterVectorFactor(InstId, DimId, Bool),
    OuterVectorFactor(InstId, HalfRange),
    IncrementNumBlockDims(InstId, DimId, Bool),
    NumBlockDims(InstId, HalfRange),
    NumNestedInst(DimId, HalfRange),
}

A decision to apply to the domain.

Variants
InstFlag(InstId, InstFlag)

cbindgen:field-names=[inst, domain]

Size(DimId, NumericSet)

cbindgen:field-names=[dim, domain]

IsThreadDim(DimId, Bool)

cbindgen:field-names=[dim, domain]

ThreadMapping(DimId, DimId, ThreadMapping)

cbindgen:field-names=[lhs, rhs, domain]

IsIterationDim(InstId, DimId, Bool)

cbindgen:field-names=[inst, dim, domain]

MemorySpace(VarId, MemorySpace)

cbindgen:field-names=[var, domain]

MemSpace(MemId, MemSpace)

cbindgen:field-names=[mem, domain]

DimKind(DimId, DimKind)

cbindgen:field-names=[dim, domain]

Order(StmtId, StmtId, Order)

cbindgen:field-names=[lhs, rhs, domain]

DimMapping(DimId, DimId, DimMapping)

cbindgen:field-names=[lhs, rhs, domain]

IncrementTilingFactor(LogicalDimId, DimId, Bool)

cbindgen:field-names=[logical, dim, domain]

TilingFactor(LogicalDimId, Range)

cbindgen:field-names=[logical, domain]

IsThreadDimClassCounter(DimId, Range)

cbindgen:field-names=[dim, domain]

NumThreads(HalfRange)

cbindgen:field-names=[domain]

NumThreadDims(HalfRange)

cbindgen:field-names=[domain]

IsIterationDimClassCounter(InstId, DimId, Range)

cbindgen:field-names=[inst, dim, domain]

IncrementMemSize(MemId, DimId, DimId, Bool)

cbindgen:field-names=[mem, lhs, rhs, domain]

MemSize(MemId, HalfRange)

cbindgen:field-names=[mem, domain]

SharedMemUsed(HalfRange)

cbindgen:field-names=[domain]

IncrementUnrollFactor(InstId, DimId, Bool)

cbindgen:field-names=[inst, dim, domain]

UnrollFactor(InstId, HalfRange)

cbindgen:field-names=[inst, domain]

IncrementInnerVectorFactor(InstId, DimId, Bool)

cbindgen:field-names=[inst, dim, domain]

InnerVectorFactor(InstId, HalfRange)

cbindgen:field-names=[inst, domain]

IncrementOuterVectorFactor(InstId, DimId, Bool)

cbindgen:field-names=[inst, dim, domain]

OuterVectorFactor(InstId, HalfRange)

cbindgen:field-names=[inst, domain]

IncrementNumBlockDims(InstId, DimId, Bool)

cbindgen:field-names=[inst, dim, domain]

NumBlockDims(InstId, HalfRange)

cbindgen:field-names=[inst, domain]

NumNestedInst(DimId, HalfRange)

cbindgen:field-names=[dim, domain]

Methods

impl Action

pub fn complement(&self, ir_instance: &Function) -> Option<Self>

Returns the action performing the complementary decision.

pub fn display<'a>(&'a self, ir_instance: &'a Function) -> DisplayAction<'a>

Returns an object that implements Display for printing the action with its associated function.

This is needed because actions only contain IDs referring to objects stored on the Function and hence can't be displayed properly without the Function.

pub fn display_domain<'a>(
&'a self,
ir_instance: &'a Function
) -> DisplayActionDomain<'a>

Returns an object that implements Display for printing an action's domain.

This is useful to manipulate something that represents an action's domain for display or debugging, since the domains have different types for different actions.