Struct telamon::search_space::DimKind

pub struct DimKind { /* fields omitted */ }

Specifies how iteration dimensions are implemented.

Methods

impl DimKind

pub const VECTOR: DimKind

VECTOR: DimKind = DimKind{bits: 1,}

The dimension is implemented by using a vector instruction.

pub const BLOCK: DimKind

BLOCK: DimKind = DimKind{bits: 2,}

The dimension is mapped to a block dimension on the device.

pub const THREAD: DimKind

THREAD: DimKind = DimKind{bits: 4,}

The dimension is mapped to a thread dimension on the device.

pub const LOOP: DimKind

LOOP: DimKind = DimKind{bits: 8,}

The dimension is implemented as a regular loop.

pub const UNROLL: DimKind

UNROLL: DimKind = DimKind{bits: 16,}

The dimension is fully unrolled.

pub const SEQUENTIAL: DimKind

SEQUENTIAL: DimKind = DimKind{bits: <DimKind>::LOOP.bits | <DimKind>::UNROLL.bits,}

The dimension is sequential.

pub const PARALLEL: DimKind

PARALLEL: DimKind = DimKind{bits:
            <DimKind>::VECTOR.bits | <DimKind>::BLOCK.bits |
                <DimKind>::THREAD.bits,}

The dimension is parallel.

pub fn list<'a>(
&self
) -> impl Iterator<Item = Self> + 'static

Lists the alternatives contained in the domain.