Struct telamon::search_space::MemSpace

#[repr(C)]

pub struct MemSpace { /* fields omitted */ }

Indicates where a memory block is located.

Methods

impl MemSpace

pub const GLOBAL: MemSpace

The block is in the device RAM.

pub const SHARED: MemSpace

The block is in the memory shared between the threads of a block.

pub const ALL: MemSpace

pub const FAILED: MemSpace

Returns the empty domain.

pub fn all() -> Self

Returns the full domain.

pub fn insert(&mut self, alternatives: Self)

Insert values in the domain.

pub fn list<'a>(
&self
) -> impl Iterator<Item = Self> + 'static

Lists the alternatives contained in the domain.

pub fn eq(&self, other: Self) -> bool

Indicates if two choices will have the same value.

pub fn neq(&self, other: Self) -> bool

Indicates if two choices cannot be equal.