Trait telamon::search_space::NumSet

pub trait NumSet {
    type Universe: ?Sized;
    fn min(&self, universe: &Self::Universe) -> u32;

    fn max(&self, universe: &Self::Universe) -> u32;

    fn into_num_set(
        &self, 
        self_universe: &Self::Universe, 
        new_universe: &[u32]
    ) -> NumericSet { ... }

    fn as_constrained(&self, universe: &Self::Universe) -> Option<u32> { ... }

    fn lt<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }

    fn gt<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }

    fn leq<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }

    fn geq<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }

    fn eq<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }

    fn neq<D: NumSet>(
        &self, 
        universe: &Self::Universe, 
        other: D, 
        other_universe: &D::Universe
    ) -> bool { ... }
}

A domain containing integers.

Associated Types

type Universe: ?Sized

Required Methods

fn min(&self, universe: &Self::Universe) -> u32

Returns the maximum value in the domain.

fn max(&self, universe: &Self::Universe) -> u32

Returns the minimum value in the domain.

Provided Methods

fn into_num_set(
&self,
self_universe: &Self::Universe,
new_universe: &[u32]
) -> NumericSet

Converts the domain into a numeric set with the given domain. Values that are not in new_universe are skipped.

fn as_constrained(&self, universe: &Self::Universe) -> Option<u32>

Returns the value of the domain, if it is constrained.

fn lt<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool

fn gt<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool

fn leq<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool

fn geq<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool

fn eq<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool

fn neq<D: NumSet>(
&self,
universe: &Self::Universe,
other: D,
other_universe: &D::Universe
) -> bool