Struct telamon::codegen::NameMap

pub struct NameMap<'a, 'b, N: Namer> { /* fields omitted */ }

Maps variables to names.

Methods

impl<'a, 'b, N: Namer> NameMap<'a, 'b, N>

pub fn new(function: &'a Function<'a>, namer: &'b mut N) -> Self

Creates a new NameMap.

pub fn namer(&self) -> &N

pub fn namer_mut(&mut self) -> &mut N

pub fn gen_name(&mut self, t: Type) -> String

Generates a variable of the given Type.

pub fn gen_loop_id(&mut self) -> u32

Generates an ID for a loop.

pub fn name(&self, operand: Operand) -> Cow<str>

pub fn name_op(&self, operand: &Operand) -> Cow<str>

Asigns a name to an operand.

pub fn name_inst(&self, inst: InstId) -> &str

Returns the name of the instruction.

pub fn indexed_inst_name(&self, inst: InstId, indexes: &[(DimId, u32)]) -> &str

pub fn indexed_op_name(
&self,
op: &Operand,
indexes: &[(DimId, u32)]
) -> Cow<str>

pub fn name_index(&self, dim_id: DimId) -> &str

Returns the name of an index.

pub fn set_current_index(&mut self, dim: &Dimension, idx: u32)

Set the current index of an unrolled dimension.

pub fn unset_current_index(&mut self, dim: &Dimension)

Unset the current index of an unrolled dimension.

pub fn name_param(&self, param: ParamValKey) -> Cow<str>

Returns the name of a variable representing a parameter.

pub fn name_param_val(&self, param: ParamValKey) -> Cow<str>

Returns the name of a variable representing a parameter value.

pub fn name_addr(&self, id: MemId) -> Cow<str>

Returns the name of the address of a memory block.

pub fn name_induction_var(&self, var: IndVarId, dim: Option<DimId>) -> Cow<str>

Assigns a name to an induction variable.

pub fn declare_size_cast(
&mut self,
size: &'a Size<'a>,
t: Type
) -> Option<String>

Declares a size cast. Returns the name of the variable only if a new variable was allcoated.

pub fn name_size(&self, size: &Size, expected_t: Type) -> Cow<str>

Assigns a name to a size.

pub fn side_effect_guard(&self) -> Option<RcStr>

Returns the side-effect guard, if any is set.

pub fn set_side_effect_guard(&mut self, guard: Option<RcStr>)

Sets the predicate to use in front of side-effect instruction.