Struct telamon::codegen::NameMap

pub struct NameMap<'a, 'b> { /* fields omitted */ }

Maps variables to names.

Methods

impl<'a, 'b> NameMap<'a, 'b>

pub fn new(function: &'a Function<'a>, namer: &'b mut Namer) -> Self

Creates a new NameMap.

pub fn gen_name(&self, t: Type) -> String

Generates a variable of the given Type.

pub fn gen_loop_id(&mut self) -> u32

Generates an ID for a loop.

pub fn name(&self, value: Value) -> Cow<str>

pub fn name_op(&self, operand: &Operand) -> Cow<str>

Asigns a name to an operand.

pub fn name_inst(&self, inst: &Instruction) -> &str

Returns the name of the instruction.

pub fn name_inst_id(&self, id: InstId, dim_map: &DimMap) -> &str

Returns the name of the instruction.

pub fn name_index(&self, dim_id: DimId) -> &str

Returns the name of an index.

pub fn set_current_index(&mut self, dim: &Dimension, idx: u32)

Set the current index of an unrolled dimension.

pub fn unset_current_index(&mut self, dim: &Dimension)

Unset the current index of an unrolled dimension.

pub fn indexed_inst_name(
&mut self,
inst: &Instruction,
dim: DimId,
idx: u32
) -> String

pub fn indexed_op_name(&mut self, op: &Operand, dim: DimId, idx: u32) -> String

pub fn name_param(&self, param: ParamValKey) -> Cow<str>

Returns the name of a variable representing a parameter.

pub fn name_param_val(&self, param: ParamValKey) -> Cow<str>

Returns the name of a variable representing a parameter value.

pub fn name_addr(&self, id: InternalId) -> Cow<str>

Returns the name of the address of a memory block.

pub fn name_induction_var(&self, var: IndVarId, dim: Option<DimId>) -> Cow<str>

Assigns a name to an induction variable.

pub fn declare_size_cast(
&mut self,
size: &'a Size<'a>,
t: Type
) -> Option<String>

Declares a size cast. Returns the name of the variable only if a new variable was allcoated.

pub fn name_size(&self, size: &Size, expected_t: Type) -> Cow<str>

Assigns a name of a value to a size.

pub fn side_effect_guard(&self) -> Option<RcStr>

Returns the side-effect guard, if any is set.

pub fn set_side_effect_guard(&mut self, guard: Option<RcStr>)

Sets the predicate to use in front of side-effect instruction.