Trait telamon::codegen::Printer

pub trait Printer {
    fn get_int(n: u32) -> String;

    fn get_float(f: f64) -> String;

    fn get_type(t: Type) -> String;

    fn print_binop(
        &mut self, 
        op: BinOp, 
        return_type: Type, 
        rounding: Rounding, 
        return_id: &str, 
        lhs: &str, 
        rhs: &str
    );

    fn print_mul(
        &mut self, 
        return_type: Type, 
        round: Rounding, 
        mul_mode: MulMode, 
        return_id: &str, 
        op1: &str, 
        op2: &str
    );

    fn print_mad(
        &mut self, 
        ret_type: Type, 
        round: Rounding, 
        mul_mode: MulMode, 
        return_id: &str, 
        mlhs: &str, 
        mrhs: &str, 
        arhs: &str
    );

    fn print_mov(&mut self, return_type: Type, return_id: &str, op: &str);

    fn print_ld(
        &mut self, 
        return_type: Type, 
        flag: InstFlag, 
        return_id: &str, 
        addr: &str
    );

    fn print_st(
        &mut self, 
        val_type: Type, 
        mem_flag: InstFlag, 
        addr: &str, 
        val: &str
    );

    fn print_cond_st(
        &mut self, 
        return_type: Type, 
        mem_flag: InstFlag, 
        cond: &str, 
        addr: &str, 
        val: &str
    );

    fn print_cast(
        &mut self, 
        from_t: Type, 
        to_t: Type, 
        round: Rounding, 
        return_id: &str, 
        op1: &str
    );

    fn print_label(&mut self, label_id: &str);

    fn print_and(&mut self, return_id: &str, op1: &str, op2: &str);

    fn print_or(&mut self, return_id: &str, op1: &str, op2: &str);

    fn print_equal(&mut self, return_id: &str, op1: &str, op2: &str);

    fn print_lt(&mut self, return_id: &str, op1: &str, op2: &str);

    fn print_gt(&mut self, return_id: &str, op1: &str, op2: &str);

    fn print_cond_jump(&mut self, label_id: &str, cond: &str);

    fn print_sync(&mut self);

    fn print_vector_inst(
        &mut self, 
        inst: &Instruction, 
        dim: &Dimension, 
        namer: &mut NameMap, 
        fun: &Function
    );

    fn mul_mode(from: Type, to: Type) -> MulMode { ... }

    fn lower_type(t: Type, fun: &Function) -> Type { ... }

    fn cfg_vec(&mut self, fun: &Function, cfgs: &[Cfg], namer: &mut NameMap) { ... }

    fn cfg<'a>(&mut self, fun: &Function, c: &Cfg<'a>, namer: &mut NameMap) { ... }

    fn parallel_induction_level(
        &mut self, 
        level: &InductionLevel, 
        namer: &NameMap
    ) { ... }

    fn enable_threads(
        &mut self, 
        fun: &Function, 
        threads: &[bool], 
        namer: &mut NameMap
    ) { ... }

    fn gen_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap
    ) { ... }

    fn standard_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap
    ) { ... }

    fn unroll_loop(
        &mut self, 
        fun: &Function, 
        dim: &Dimension, 
        cfgs: &[Cfg], 
        namer: &mut NameMap
    ) { ... }

    fn privatise_global_block(
        &mut self, 
        block: &InternalMemBlock, 
        namer: &mut NameMap, 
        fun: &Function
    ) { ... }

    fn inst(&mut self, inst: &Instruction, namer: &mut NameMap, fun: &Function) { ... }
}

Required Methods

fn get_int(n: u32) -> String

Get a proper string representation of an integer in target language

fn get_float(f: f64) -> String

Get a proper string representation of an float in target language

fn get_type(t: Type) -> String

Print a type in the backend

fn print_binop(
&mut self,
op: BinOp,
return_type: Type,
rounding: Rounding,
return_id: &str,
lhs: &str,
rhs: &str
)

Print return_id = lhs op rhs

fn print_mul(
&mut self,
return_type: Type,
round: Rounding,
mul_mode: MulMode,
return_id: &str,
op1: &str,
op2: &str
)

Print return_id = op1 * op2

fn print_mad(
&mut self,
ret_type: Type,
round: Rounding,
mul_mode: MulMode,
return_id: &str,
mlhs: &str,
mrhs: &str,
arhs: &str
)

Print return_id = mlhs * mrhs + arhs

fn print_mov(&mut self, return_type: Type, return_id: &str, op: &str)

Print return_id = op

fn print_ld(
&mut self,
return_type: Type,
flag: InstFlag,
return_id: &str,
addr: &str
)

Print return_id = load [addr]

fn print_st(
&mut self,
val_type: Type,
mem_flag: InstFlag,
addr: &str,
val: &str
)

Print store val [addr]

fn print_cond_st(
&mut self,
return_type: Type,
mem_flag: InstFlag,
cond: &str,
addr: &str,
val: &str
)

Print if (cond) store val [addr]

fn print_cast(
&mut self,
from_t: Type,
to_t: Type,
round: Rounding,
return_id: &str,
op1: &str
)

Print return_id = (val_type) val

fn print_label(&mut self, label_id: &str)

print a label where to jump

fn print_and(&mut self, return_id: &str, op1: &str, op2: &str)

Print return_id = op1 && op2

fn print_or(&mut self, return_id: &str, op1: &str, op2: &str)

Print return_id = op1 || op2

fn print_equal(&mut self, return_id: &str, op1: &str, op2: &str)

Print return_id = op1 == op2

fn print_lt(&mut self, return_id: &str, op1: &str, op2: &str)

Print return_id = op1 < op2

fn print_gt(&mut self, return_id: &str, op1: &str, op2: &str)

Print return_id = op1 > op2

fn print_cond_jump(&mut self, label_id: &str, cond: &str)

Print if (cond) jump label(label_id)

fn print_sync(&mut self)

Print wait on all threads

fn print_vector_inst(
&mut self,
inst: &Instruction,
dim: &Dimension,
namer: &mut NameMap,
fun: &Function
)

Provided Methods

fn mul_mode(from: Type, to: Type) -> MulMode

fn lower_type(t: Type, fun: &Function) -> Type

fn cfg_vec(&mut self, fun: &Function, cfgs: &[Cfg], namer: &mut NameMap)

fn cfg<'a>(&mut self, fun: &Function, c: &Cfg<'a>, namer: &mut NameMap)

Prints a cfg.

fn parallel_induction_level(&mut self, level: &InductionLevel, namer: &NameMap)

Prints a multiplicative induction var level.

fn enable_threads(
&mut self,
fun: &Function,
threads: &[bool],
namer: &mut NameMap
)

Change the side-effect guards so that only the specified threads are enabled.

fn gen_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap
)

Prints a Loop

fn standard_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap
)

Prints a classic loop - that is, a sequential loop with an index and a jump to the beginning at the end of the block

fn unroll_loop(
&mut self,
fun: &Function,
dim: &Dimension,
cfgs: &[Cfg],
namer: &mut NameMap
)

Prints an unroll loop - loop without jumps

fn privatise_global_block(
&mut self,
block: &InternalMemBlock,
namer: &mut NameMap,
fun: &Function
)

fn inst(&mut self, inst: &Instruction, namer: &mut NameMap, fun: &Function)

Prints an instruction.