Trait telamon::device::Device

pub trait Device: Sync {
    fn print(&self, function: &Function, out: &mut Write);

    fn check_type(&self, t: Type) -> Result<(), TypeError>;

    fn max_block_dims(&self) -> u32;

    fn max_inner_block_size(&self) -> u32;

    fn max_threads(&self) -> u32;

    fn max_unrolling(&self) -> u32;

    fn vectorization_factors(&self, dim: &Dimension, op: &Operator) -> &[u32];

    fn shared_mem(&self) -> u32;

    fn supports_nc_access(&self) -> bool;

    fn supports_l1_access(&self) -> bool;

    fn supports_l2_access(&self) -> bool;

    fn name(&self) -> &str;

    fn hw_pressure(
        &self, 
        space: &SearchSpace, 
        dim_sizes: &HashMap<DimId, Range>, 
        nesting: &HashMap<StmtId, Nesting>, 
        bb: &Statement, 
        ctx: &Context
    ) -> HwPressure;

    fn loop_iter_pressure(&self, kind: DimKind) -> (HwPressure, HwPressure);

    fn thread_rates(&self) -> HwPressure;

    fn block_rates(&self) -> HwPressure;

    fn total_rates(&self) -> HwPressure;

    fn bottlenecks(&self) -> &[&'static str];

    fn block_parallelism(&self, space: &SearchSpace) -> u32;

    fn additive_indvar_pressure(&self, t: &Type) -> HwPressure;

    fn multiplicative_indvar_pressure(&self, t: &Type) -> HwPressure;

    fn add_block_overhead(
        &self, 
        max_active_threads: FactorRange, 
        max_threads: FactorRange, 
        predication_factor: Range, 
        pressure: &mut HwPressure
    );

    fn lower_type(&self, t: Type, space: &SearchSpace) -> Option<Type>;

    fn gen_code(&self, implementation: &SearchSpace, out: &mut Write) { ... }
}

Holds the specifications of a target.

Required Methods

fn print(&self, function: &Function, out: &mut Write)

Prints the code corresponding to a device Function.

fn check_type(&self, t: Type) -> Result<(), TypeError>

Indicates if a Type can be implemented on the device.

fn max_block_dims(&self) -> u32

Returns the maximal number of block dimensions.

fn max_inner_block_size(&self) -> u32

The maximal size inner block dimensions can have.

fn max_threads(&self) -> u32

Returns the maximal number of threads.

fn max_unrolling(&self) -> u32

Returns the maximal unrolling factor.

fn vectorization_factors(&self, dim: &Dimension, op: &Operator) -> &[u32]

Indicates the valid vectorization factors for the given operator, on the given dimension.

fn shared_mem(&self) -> u32

Returns the amount of shared memory available for each thread block.

fn supports_nc_access(&self) -> bool

Indicates if the device supports non-coherent memory accesses.

fn supports_l1_access(&self) -> bool

Indicates if the device supports L1 for global memory accesses.

fn supports_l2_access(&self) -> bool

Indicates if the device supports L2 for global memory accesses.

fn name(&self) -> &str

Returns the name of the device.

fn hw_pressure(
&self,
space: &SearchSpace,
dim_sizes: &HashMap<DimId, Range>,
nesting: &HashMap<StmtId, Nesting>,
bb: &Statement,
ctx: &Context
) -> HwPressure

Returns the pressure cause by a Statement. For a dimension, returns the pressure for the full loop execution.

fn loop_iter_pressure(&self, kind: DimKind) -> (HwPressure, HwPressure)

Returns the pressure produced by a single iteration of a loop and the latency overhead of iterations.

fn thread_rates(&self) -> HwPressure

Returns the processing rates of a single thread, in units/ns

fn block_rates(&self) -> HwPressure

Returns the processing rates of a single block, in units/ns.

fn total_rates(&self) -> HwPressure

Returns the processing rates of the whole accelerator un units/ns.

fn bottlenecks(&self) -> &[&'static str]

Returns the names of potential bottlenecks.

fn block_parallelism(&self, space: &SearchSpace) -> u32

Returns the number of blocks that can be executed in parallel on the device.

fn additive_indvar_pressure(&self, t: &Type) -> HwPressure

Returns the pressure caused by an additive induction variable level.

fn multiplicative_indvar_pressure(&self, t: &Type) -> HwPressure

Returns the pressure caused by a multiplicative induction variable level.

fn add_block_overhead(
&self,
max_active_threads: FactorRange,
max_threads: FactorRange,
predication_factor: Range,
pressure: &mut HwPressure
)

Adds the overhead (per instance) due to partial wraps and predicated dimensions to the pressure. If the instruction is not predicated, predicated_dims_size should be 1.

fn lower_type(&self, t: Type, space: &SearchSpace) -> Option<Type>

Lowers a type using the memory space information. Returns None if some information is not yet specified.

Provided Methods

fn gen_code(&self, implementation: &SearchSpace, out: &mut Write)

Builds and outputs a constrained IR instance.