Trait telamon::device::ScalarArgument

pub unsafe trait ScalarArgument: Sync + Send + Copy + PartialEq + Display + 'static {
    fn t() -> Type;

    fn raw_ptr(&self) -> *const c_void;

    fn as_operand<L>(&self) -> Operand<'static, L>;

    fn gen_random<R: Rng>(_: &mut R) -> Self;

    fn as_size(&self) -> Option<u32> { ... }
}

Represents a value that can be used as a Function argument. Must ensures the type is a scalar and does not contains any reference.

Required Methods

fn t() -> Type

Returns the type of the argument.

fn raw_ptr(&self) -> *const c_void

Returns a raw pointer to the object.

fn as_operand<L>(&self) -> Operand<'static, L>

Returns an operand holding the argument value as a constant.

fn gen_random<R: Rng>(_: &mut R) -> Self

Generates a random instance of the argument type.

Provided Methods

fn as_size(&self) -> Option<u32>

Returns the argument interpreted as an iteration dimension size, if applicable.