Module telamon::explorer::config

Defines a structure to store the configuration of the exploration. The configuration is read from the Setting.toml file if it exists. Some parameters can be overridden from the command line.

Structs

BanditConfig
Configuration parameters specific to the multi-armed bandit algorithm.

Config
Stores the configuration of the exploration.

Enums

NewNodeOrder
Indicates how to choose between nodes of the search tree when no children have been evaluated.

OldNodeOrder
Indicates how to choose between nodes of the search tree with at least one descendent evaluated.

SearchAlgorithm
Exploration algorithm to use.