Struct telamon::explorer::config::Config

pub struct Config {
    pub log_file: String,
    pub num_workers: usize,
    pub algorithm: SearchAlgorithm,
    pub stop_bound: Option<f64>,
    pub timeout: Option<u64>,
    pub distance_to_best: Option<f64>,
}

Stores the configuration of the exploration.

Fields
log_file: String

Name of the file in wich to store the logs.

num_workers: usize

Number of exploration threads.

algorithm: SearchAlgorithm

Exploration algorithm to use.

stop_bound: Option<f64>

Indicates the search must be stopped if a candidate with an execution time better than the bound (in ns) is found.

timeout: Option<u64>

Indicates the search must be stopped after the given number of minutes.

distance_to_best: Option<f64>

A percentage cut indicate that we only care to find a candidate that is in a certain range above the best Therefore, if cut_under is 20%, we can discard any candidate whose bound is above 80% of the current best.

Methods

impl Config

pub fn read() -> Self

Reads the configuration from the "Settings.toml" file and from the command line.

pub fn read_from_file() -> Self

Extract the configuration from the configuration file, if any.