Struct telamon::helper::Builder

pub struct Builder<'a> { /* fields omitted */ }

Helper to build a Function.

Methods

impl<'a> Builder<'a>

pub fn new(signature: &'a Signature, device: &'a dyn Device) -> Builder<'a>

Creates a new Builder for a Function with the given signature.

pub fn get(self) -> SearchSpace<'a>

Returns the function created by the builder

pub fn get_clone(&self) -> SearchSpace<'a>

Returns the function created by the builder

pub fn binop<'b: 'a>(
&mut self,
op: BinOp,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>
) -> InstId

Creates a binary operator.

pub fn add<'b: 'a>(
&mut self,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>
) -> InstId

Adds an Add instruction to the fuction.

pub fn sub<'b: 'a>(
&mut self,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>
) -> InstId

Adds a Sub instruction to the function.

pub fn mul<'b: 'a>(
&mut self,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>
) -> InstId

Adds a Mul instruction to the function. Defaults to low mode.

pub fn mul_ex<'b: 'a>(
&mut self,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>,
t: Type
) -> InstId

Adds a 'Mul` instruction with a wide mode to the function.

pub fn mad<'b: 'a>(
&mut self,
mul_lhs: &dyn AutoOperand<'b>,
mul_rhs: &dyn AutoOperand<'b>,
add_rhs: &dyn AutoOperand<'b>
) -> InstId

Adds a Mad or Fma instruction to the function. Defaults to low or wide mode depending on the operand types.

pub fn div<'b: 'a>(
&mut self,
lhs: &dyn AutoOperand<'b>,
rhs: &dyn AutoOperand<'b>
) -> InstId

Adds a Div instruction to the fuction.

pub fn mov<'b: 'a>(&mut self, arg: &dyn AutoOperand<'b>) -> InstId

Adds a Mov instruction to the function.

pub fn ld<'b: 'a>(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a coherent load from global memory instruction to the function.

pub fn ld_nc<'b: 'a>(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a non-coherent load from global memory instruction to the function.

pub fn ld_ex<'b: 'a>(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand<'b>,
pattern: AccessPattern<'a>,
flags: InstFlag
) -> InstId

Adds a load instruction with the given flags and memory block.

pub fn st<'b: 'a>(
&mut self,
addr: &dyn AutoOperand<'b>,
val: &dyn AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a store instruction.

pub fn st_ex<'b: 'a>(
&mut self,
addr: &dyn AutoOperand<'b>,
val: &dyn AutoOperand<'b>,
side_effect: bool,
pattern: AccessPattern<'a>,
flags: InstFlag
) -> InstId

Adds a store instruction with the given flags and memory block.

pub fn cast<'b: 'a>(&mut self, val: &dyn AutoOperand<'b>, t: Type) -> InstId

Adds a cast instruction to the given type.

pub fn order(
&mut self,
lhs: &dyn MetaStatement,
rhs: &dyn MetaStatement,
order: Order
)

Restricts the order between two basic blocks. Does not restricts LINK and NPACK flags.

pub fn get_inst_variable(&mut self, inst_id: InstId) -> VarId

Returns the variable holding the result of an instruction. Creates it if necessary.

pub fn create_last_variable(
&mut self,
var: VarId,
logical_dims: &[&LogicalDim]
) -> VarId

Creates a new variable that takes the last value of another variable produced in a loop nest.

pub fn create_dim_map_variable(
&mut self,
var: VarId,
logical_mapping: &[(&LogicalDim, &LogicalDim)]
) -> VarId

Creates a new variable that takes point-to-point the value of another variable, in another loop nest.

pub fn action(&mut self, action: Action)

Applies an action on the function.

pub fn open_dim(&mut self, size: Size<'a>) -> LogicalDim

Opens a new dimension.

pub fn open_dim_ex(&mut self, size: Size<'a>, kind: DimKind) -> LogicalDim

Opens a nest of new dimension with the given kinds and sizes.

pub fn open_tiled_dim(
&mut self,
size: Size<'a>,
tiling_pattern: TilingPattern
) -> LogicalDim

Open multiple dimensions to represent a tiled dimension.

pub fn open_mapped_dim(&mut self, old_dim: &LogicalDim) -> LogicalDim

Opens a new dimension mapped to an existing one.

The size of the new dim is inherited from the mapped dim. The dimension mapped to is closed if needed.

pub fn reopen_dim(&mut self, dim: &LogicalDim)

Opens an existing dimension.

pub fn reopen_mapped_dim(&mut self, dim: &LogicalDim, mapped_to: &LogicalDim)

Opens an existing dimension and maps it to another one. The dimension mapped to is closed if needed.

pub fn close_dim(&mut self, dims: &LogicalDim)

Closes a dimension.

pub fn cst_size(&self, size: u32) -> Size<'a>

Returns a constant size.

pub fn param_size(&self, param: &str, max_size: u32) -> Size<'a>

Returns a parameter size.

pub fn allocate_shared(&mut self, size: u32) -> MemId

Allocates a memory block in shared memory.

pub fn allocate(&mut self, size: u32, private: bool) -> MemId

Allocates a memory block.

pub fn tensor_access(
&mut self,
addr: &dyn AutoOperand<'a>,
mem_id: Option<MemId>,
t: Type,
dims: &[&LogicalDim]
) -> (IndVarId, AccessPattern<'a>)

Builds both an induction variable for a tensor memory access and the corresponding access pattern.

pub fn tensor_access_pattern(
&self,
mem: Option<MemId>,
increments: Vec<(&LogicalDim, Size<'a>)>
) -> AccessPattern<'a>

Generates the access pattern corresponding to accessing a tensor of the given type.

pub fn induction_var(
&mut self,
base: &dyn AutoOperand<'a>,
dims: Vec<(&LogicalDim, Size<'a>)>
) -> IndVarId

Builds an induction variable.

pub fn dim_map(
&self,
base: InstId,
dim_map: &[(&LogicalDim, &LogicalDim)],
scope: DimMapScope<()>
) -> Operand<'a, ()>

Creates a dim-map operand.

pub fn find_param(&self, param: &str) -> &'a Parameter

Finds a paramter given its name.