Struct telamon::helper::Builder

pub struct Builder { /* fields omitted */ }

Helper to build a Function.

Methods

impl Builder

pub fn new(signature: Arc<Signature>, device: Arc<dyn Device>) -> Builder

Creates a new Builder for a Function with the given signature.

pub fn get(self) -> SearchSpace

Returns the function created by the builder

pub fn get_clone(&self) -> SearchSpace

Returns the function created by the builder

pub fn binop(
&mut self,
op: BinOp,
lhs: &dyn AutoOperand,
rhs: &dyn AutoOperand
) -> InstId

Creates a binary operator.

pub fn add(&mut self, lhs: &dyn AutoOperand, rhs: &dyn AutoOperand) -> InstId

Adds an Add instruction to the fuction.

pub fn sub(&mut self, lhs: &dyn AutoOperand, rhs: &dyn AutoOperand) -> InstId

Adds a Sub instruction to the function.

pub fn mul(&mut self, lhs: &dyn AutoOperand, rhs: &dyn AutoOperand) -> InstId

Adds a Mul instruction to the function. Defaults to low mode.

pub fn mul_ex(
&mut self,
lhs: &dyn AutoOperand,
rhs: &dyn AutoOperand,
t: Type
) -> InstId

Adds a 'Mul` instruction with a wide mode to the function.

pub fn mad(
&mut self,
mul_lhs: &dyn AutoOperand,
mul_rhs: &dyn AutoOperand,
add_rhs: &dyn AutoOperand
) -> InstId

Adds a Mad or Fma instruction to the function. Defaults to low or wide mode depending on the operand types.

pub fn max(&mut self, lhs: &dyn AutoOperand, rhs: &dyn AutoOperand) -> InstId

Adds a Max instruction to the fuction.

pub fn div(&mut self, lhs: &dyn AutoOperand, rhs: &dyn AutoOperand) -> InstId

Adds a Div instruction to the fuction.

pub fn mov(&mut self, arg: &dyn AutoOperand) -> InstId

Adds a Mov instruction to the function.

pub fn exp(&mut self, arg: &dyn AutoOperand) -> InstId

Adds an Exp instruction to the function.

pub fn ld(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand,
pattern: AccessPattern
) -> InstId

Adds a coherent load from global memory instruction to the function.

pub fn ld_nc(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand,
pattern: AccessPattern
) -> InstId

Adds a non-coherent load from global memory instruction to the function.

pub fn ld_ex(
&mut self,
ret_type: Type,
addr: &dyn AutoOperand,
pattern: AccessPattern,
flags: InstFlag
) -> InstId

Adds a load instruction with the given flags and memory block.

pub fn st(
&mut self,
addr: &dyn AutoOperand,
val: &dyn AutoOperand,
pattern: AccessPattern
) -> InstId

Adds a store instruction.

pub fn st_ex(
&mut self,
addr: &dyn AutoOperand,
val: &dyn AutoOperand,
side_effect: bool,
pattern: AccessPattern,
flags: InstFlag
) -> InstId

Adds a store instruction with the given flags and memory block.

pub fn cast(&mut self, val: &dyn AutoOperand, t: Type) -> InstId

Adds a cast instruction to the given type.

pub fn order(
&mut self,
lhs: &dyn MetaStatement,
rhs: &dyn MetaStatement,
order: Order
)

Restricts the order between two basic blocks. Does not restricts LINK and NPACK flags.

pub fn get_inst_variable(&mut self, inst_id: InstId) -> VarId

Returns the variable holding the result of an instruction. Creates it if necessary.

pub fn create_last_variable(
&mut self,
var: VarId,
logical_dims: &[&LogicalDim]
) -> VarId

Creates a new variable that takes the last value of another variable produced in a loop nest.

pub fn create_dim_map_variable(
&mut self,
var: VarId,
logical_mapping: &[(&LogicalDim, &LogicalDim)]
) -> VarId

Creates a new variable that takes point-to-point the value of another variable, in another loop nest.

pub fn action(&mut self, action: Action)

Applies an action on the function.

pub fn open_dim(&mut self, size: Size) -> LogicalDim

Opens a new dimension.

pub fn open_dim_ex(&mut self, size: Size, kind: DimKind) -> LogicalDim

Opens a nest of new dimension with the given kinds and sizes.

pub fn open_tiled_dim(
&mut self,
size: Size,
tiling_pattern: TilingPattern
) -> LogicalDim

Open multiple dimensions to represent a tiled dimension.

pub fn open_mapped_dim(&mut self, old_dim: &LogicalDim) -> LogicalDim

Opens a new dimension mapped to an existing one.

The size of the new dim is inherited from the mapped dim. The dimension mapped to is closed if needed.

pub fn reopen_dim(&mut self, dim: &LogicalDim)

Opens an existing dimension.

pub fn reopen_mapped_dim(&mut self, dim: &LogicalDim, mapped_to: &LogicalDim)

Opens an existing dimension and maps it to another one. The dimension mapped to is closed if needed.

pub fn close_dim(&mut self, dims: &LogicalDim)

Closes a dimension.

pub fn cst_size(&self, size: u32) -> Size

Returns a constant size.

pub fn param_size(&self, param: &str, max_size: u32) -> Size

Returns a parameter size.

pub fn allocate_shared(&mut self, size: u32) -> MemId

Allocates a memory block in shared memory.

pub fn allocate(&mut self, size: u32, private: bool) -> MemId

Allocates a memory block.

pub fn tensor_access(
&mut self,
addr: &dyn AutoOperand,
mem_id: Option<MemId>,
t: Type,
dims: &[&LogicalDim]
) -> (IndVarId, AccessPattern)

Builds both an induction variable for a tensor memory access and the corresponding access pattern.

pub fn tensor_access_pattern(
&self,
mem: Option<MemId>,
increments: Vec<(&LogicalDim, Size)>
) -> AccessPattern

Generates the access pattern corresponding to accessing a tensor of the given type.

pub fn induction_var(
&mut self,
base: &dyn AutoOperand,
dims: Vec<(&LogicalDim, Size)>
) -> IndVarId

Builds an induction variable.

pub fn dim_map(
&self,
base: InstId,
dim_map: &[(&LogicalDim, &LogicalDim)],
scope: DimMapScope<()>
) -> Operand<()>

Creates a dim-map operand.

pub fn find_param(&self, param: &str) -> &Arc<Parameter>

Finds a paramter given its name.

pub fn function(&self) -> &Function<()>

Returns a reference to the function being built.