Struct telamon::helper::Builder

pub struct Builder<'a> { /* fields omitted */ }

Helper to build a Function.

Methods

impl<'a> Builder<'a>

pub fn new(signature: &'a Signature, device: &'a Device) -> Builder<'a>

Creates a new Builder for a Function with the given signature.

pub fn get(self) -> SearchSpace<'a>

Returns the function created by the builder

pub fn get_clone(&self) -> SearchSpace<'a>

Returns the function created by the builder

pub fn binop<'b: 'a>(
&mut self,
op: BinOp,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>
) -> InstId

Creates a binary operator.

pub fn add<'b: 'a>(
&mut self,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>
) -> InstId

Adds an Add instruction to the fuction.

pub fn sub<'b: 'a>(
&mut self,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>
) -> InstId

Adds a Sub instruction to the function.

pub fn mul<'b: 'a>(
&mut self,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>
) -> InstId

Adds a Mul instruction to the function. Defaults to low mode.

pub fn mul_ex<'b: 'a>(
&mut self,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>,
t: Type
) -> InstId

Adds a 'Mul` instruction with a wide mode to the function.

pub fn mad<'b: 'a>(
&mut self,
mul_lhs: &AutoOperand<'b>,
mul_rhs: &AutoOperand<'b>,
add_rhs: &AutoOperand<'b>
) -> InstId

Adds a Mad or Fma instruction to the function. Defaults to low or wide mode depending on the operand types.

pub fn div<'b: 'a>(
&mut self,
lhs: &AutoOperand<'b>,
rhs: &AutoOperand<'b>
) -> InstId

Adds a Div instruction to the fuction.

pub fn mov<'b: 'a>(&mut self, arg: &AutoOperand<'b>) -> InstId

Adds a Mov instruction to the function.

pub fn ld<'b: 'a>(
&mut self,
ret_type: Type,
addr: &AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a coherent load from global memory instruction to the function.

pub fn ld_nc<'b: 'a>(
&mut self,
ret_type: Type,
addr: &AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a non-coherent load from global memory instruction to the function.

pub fn ld_ex<'b: 'a>(
&mut self,
ret_type: Type,
addr: &AutoOperand<'b>,
pattern: AccessPattern<'a>,
flags: InstFlag
) -> InstId

Adds a load instruction with the given flags and memory block.

pub fn st<'b: 'a>(
&mut self,
addr: &AutoOperand<'b>,
val: &AutoOperand<'b>,
pattern: AccessPattern<'a>
) -> InstId

Adds a store instruction.

pub fn st_ex<'b: 'a>(
&mut self,
addr: &AutoOperand<'b>,
val: &AutoOperand<'b>,
side_effect: bool,
pattern: AccessPattern<'a>,
flags: InstFlag
) -> InstId

Adds a store instruction with the given flags and memory block.

pub fn cast<'b: 'a>(&mut self, val: &AutoOperand<'b>, t: Type) -> InstId

Adds a cast instruction to the given type.

pub fn type_of<'b: 'a>(&self, op: &AutoOperand<'b>) -> Type

Returns the type of an operand.

pub fn order(
&mut self,
lhs: &MetaBasicBlock,
rhs: &MetaBasicBlock,
order: Order
)

Restricts the order between two basic blocks. Does not restricts LINK and NPACK flags.

pub fn tensor_access(
&mut self,
addr: &AutoOperand<'a>,
mem: MemId,
t: &Type,
dims: &[&MetaDimension]
) -> (IndVarId, AccessPattern<'a>)

Builds both an induction variable for a tensor memory access and the corresponding access pattern.

pub fn action(&mut self, action: Action)

Applies an action on the function.

pub fn open_dim(&mut self, size: Size<'a>) -> DimId

Opens a new dimension.

pub fn open_dim_ex(&mut self, size: Size<'a>, kind: DimKind) -> DimId

Opens a nest of new dimension with the given kinds and sizes.

pub fn open_tiled_dim(&mut self, size: Size<'a>, tiling: &[u32]) -> DimGroup

Open multiple dimensions to represent a tiled dimension.

pub fn open_mapped_dim(&mut self, old_dim: &MetaDimension) -> DimGroup

Opens a new dimension mapped to an existing one.

The size of the new dim is inherited from the mapped dim. The dimension mapped to is closed if needed.

pub fn reopen_dim(&mut self, dim: &MetaDimension)

Opens an existing dimension.

pub fn reopen_mapped_dim(
&mut self,
dim: &MetaDimension,
mapped_to: &MetaDimension
)

Opens an existing dimension and maps it to another one. The dimension mapped to is closed if needed.

pub fn close_dim(&mut self, dims: &MetaDimension)

Closes a dimension.

pub fn cst_size(&self, size: u32) -> Size<'a>

Returns a constant size.

pub fn param_size(&self, param: &str) -> Size<'a>

Returns a parameter size.

pub fn tile_size(&self, param: &str, chunk_size: u32) -> Size<'a>

Returns a tiled size.

pub fn size(&self, params: &[&str], dividend: u32, divisor: u32) -> Size<'a>

Returns a size from the given parameters, dividend and divisor.

pub fn allocate_shared(&mut self, size: u32) -> InternalId

Allocates a memory block in shared memory.

pub fn allocate(&mut self, size: u32, private: bool) -> InternalId

Allocates a memory block.

pub fn unknown_access_pattern(&self, mem: MemId) -> AccessPattern<'static>

Generates an access paterns with all the strides unknown on the opened dimensions.

pub fn tensor_access_pattern(
&self,
mem: MemId,
t: &Type,
dims: &[&MetaDimension]
) -> AccessPattern<'a>

Generates the access pattern corresponding to accessing a tensor of the given type. The data is assumed to be laid out contiguously in the order given by dimensions. The last dimension is the major order.

pub fn induction_var(
&mut self,
base: &AutoOperand<'a>,
dims: Vec<(DimId, Size<'a>)>
) -> IndVarId

Builds an induction variable.

pub fn dim_map(
&self,
base: InstId,
dim_map: &[(&MetaDimension, &MetaDimension)],
scope: DimMapScope<()>
) -> Operand<'a, ()>

Creates a dim-map operand.

pub fn find_param(&self, param: &str) -> &'a Parameter

Finds a paramter given its name.