Enum telamon::ir::Operand

pub enum Operand<'a, L = LoweringMap> {
    Int(BigInt, u16),
    Float(Ratio<BigInt>, u16),
    Inst(InstId, Type, DimMap, DimMapScope<L>),
    Index(DimId),
    Param(&'a Parameter),
    Addr(InternalId),
    Reduce(InstId, Type, DimMap, Vec<DimId>),
    InductionVar(IndVarId, Type),
}

Represents an instruction operand.

Variants
Int(BigInt, u16)

An integer constant, on a given number of bits.

Float(Ratio<BigInt>, u16)

A float constant, on a given number of bits.

Inst(InstId, Type, DimMap, DimMapScope<L>)

A value produced by an instruction. The boolean indicates if the DimMap can be lowered.

Index(DimId)

The current index in a loop.

Param(&'a Parameter)

A parameter of the function.

Addr(InternalId)

The address of a memory block.

Reduce(InstId, Type, DimMap, Vec<DimId>)

The value of the current instruction at a previous iteration.

InductionVar(IndVarId, Type)

A variable increased by a fixed amount at every step of some loops.

Methods

impl<'a, L> Operand<'a, L>

pub fn t(&self) -> Type

Returns the type of the Operand.

pub fn new_inst(
inst: &Instruction<L>,
dim_map: DimMap,
scope: DimMapScope<L>
) -> Self

Create an operand from an instruction.

pub fn new_reduce(
init: &Instruction<L>,
dim_map: DimMap,
dims: Vec<DimId>
) -> Self

Creates a reduce operand from an instruction and a set of dimensions to reduce on.

pub fn new_int(val: BigInt, len: u16) -> Self

Creates a new Int operand and checks its number of bits.

pub fn new_float(val: Ratio<BigInt>, len: u16) -> Self

Creates a new Float operand.

pub fn merge_dims(&mut self, lhs: DimId, rhs: DimId)

Renames a basic block id.

pub fn should_lower_map(&self, lhs: DimId, rhs: DimId) -> bool

Indicates if a DimMap should be lowered if lhs and rhs are not mapped.

pub fn as_reduction(&self) -> Option<(InstId, &DimMap, &[DimId])>

If the operand is a reduction, returns the instruction initializing the reduction.

pub fn is_constant(&self) -> bool

Indicates if the operand stays constant during the execution.

impl<'a> Operand<'a, ()>

pub fn freeze(self, cnt: &mut Counter) -> Operand<'a>