Struct telamon::ir::mem::BlockMap

pub struct BlockMap { /* fields omitted */ }

Holds the blocks of memory to allocate on the device.

Methods

impl BlockMap

pub fn new(num_external: u32) -> Self

Creates a new BlocksMap.

pub fn num_internal_blocks(&self) -> usize

pub fn alloc_block(
&mut self,
base_size: u32,
maybe_mapped: Option<DimMap>
) -> InternalId

Allocates a new Block with the given type and sizes. Must call not merged on the dimensions that cannot be merged upon creation.

pub fn expand_internal_blocks_to(&mut self, capacity: usize)

pub fn set_lazy_tmp<IT>(&mut self, id: InternalId, t: Type, dims: IT) where
IT: Iterator<Item = (DimId, DimId)>,

Inserts a new temporary memory. Must be inserted before not_merged is called on dimensions.

pub fn register_use(&mut self, mem: MemId, inst: InstId)

Registers a use of a memory block by an instruction.

pub fn block(&self, id: MemId) -> &Block

Returns a block given its Id.

pub fn block_mut(&mut self, id: MemId) -> &mut Block

Returns a block given its Id.

pub fn internal_block(&self, id: InternalId) -> &InternalBlock

Returns the internal block given its ID.

pub fn internal_blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b InternalBlock>

Retuns the list of internal blocks.

pub fn blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b Block>

Returns the list of memory blocks.

pub fn merge_dims(&mut self, lhs: DimId, rhs: DimId) -> Vec<InternalId>

Rename a basic block. Returns the lyaouts to lower.

pub fn not_merged(&mut self, lhs_dim: &Dimension, rhs: DimId) -> Vec<InternalId>

Registers that two dimensions may not be merged. Returns a list of dimensions removed from the memory blocks and a list of layouts to lower.

pub fn lower_layout(&mut self, id: InternalId) -> Vec<(DimId, DimId)>

Lowers a fully defined layout. Returns the mapping of dimensions.