Struct telamon::ir::Dimension

pub struct Dimension<'a> { /* fields omitted */ }

Represents an iteration dimension.

Methods

impl<'a> Dimension<'a>

pub fn new(size: Size, id: DimId) -> Result<Dimension, Error>

Creates a new dimension.

pub fn with_same_size(id: DimId, other: &Self) -> Self

Creates a new dimension with the same size as an existing one.

pub fn size(&self) -> &Size<'a>

Retruns the size of the dimension.

pub fn possible_sizes(&self) -> Option<&[u32]>

Returns the values the size can take, if it is statically known.

pub fn id(&self) -> DimId

Returns the id of the dimension.

pub fn iterated<'b>(
&'b self
) -> impl Iterator<Item = InstId> + 'b

Returns the constructs iterated along this dimension.

pub fn add_iterated(&mut self, inst: InstId)

Adds a bb that is iterated along self.

pub fn is_thread_dim(&self) -> bool

Indicates if the dimension is a thread dimension.

pub fn set_thread_dim(&mut self)

Sets the dimension as a thread dimension.