Struct telamon::ir::Function

pub struct Function<'a, L = LoweringMap> { /* fields omitted */ }

Describes a function and the set of its possible implementation.

The type parameter L indicates if the function is fozen (L = ir::LoweringMap) or not (L = ()). A frozen function cannot have any more IDs allocated. We use this to know the exact list of potential decisions before we run the exploration.

Methods

impl<'a, L> Function<'a, L>

pub fn new(signature: &'a Signature, device: &'a dyn Device) -> Self

Creates a new function.

pub fn signature(&self) -> &'a Signature

Returns the function signature.

pub fn device(&self) -> &'a dyn Device

Returns the device the function is compiled for.

pub fn add_ind_var(&mut self, ind_var: InductionVar<'a, L>) -> IndVarId

Adds an induction variable.

pub fn insts<'b>(
&'b self
) -> impl Iterator<Item = &'b Instruction<'a, L>>

Returns the list of instructions of the function.

pub fn dims<'b>(
&'b self
) -> impl Iterator<Item = &'b Dimension<'a, L>> + Clone

Returns the list of dimensions of the function.

pub fn logical_dims(
&self
) -> impl Iterator<Item = &LogicalDim<'a>>

Returns the list of logical dimensions.

pub fn static_dims<'b>(
&'b self
) -> impl Iterator<Item = &'b Dimension<'a, L>>

Returns the list of stastic dimensions in the function.

pub fn variables(
&self
) -> impl Iterator<Item = &Variable>

pub fn thread_dims(
&self
) -> impl Iterator<Item = &Dimension<'a, L>>

Returns the list of thread dimensions.

pub fn inst(&self, id: InstId) -> &Instruction<'a, L>

Returns an instruction given its id.

pub fn dim(&self, id: DimId) -> &Dimension<'a, L>

Retuns a dimension given its id.

pub fn statement(&self, id: StmtId) -> &dyn Statement<'a, L>

Returns a Statement given its id.

pub fn statements<'b>(
&'b self
) -> impl Iterator<Item = &'b dyn Statement<'a, L>>

Lists all Statements.

pub fn logical_dim(&self, id: LogicalDimId) -> &LogicalDim<'a>

Retrives a logical dimension given its ID.

pub fn variable(&self, id: VarId) -> &Variable

Returns a Variable given its id.

pub fn add_variable(&mut self, def: VarDef) -> Result<VarId, Error>

Adds a variable to the function. Also register its definition into the relevant instruction

pub fn mem_blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b Block>

Returns the list of memory blocks. The block with id i is in i-th position.

pub fn mem_insts<'b>(
&'b self
) -> impl Iterator<Item = &'b Instruction<'a, L>> + 'b

Iterates over memory instructions.

pub fn mem_block(&self, id: MemId) -> &Block

Returns a memory block given its id.

pub fn induction_var(&self, id: IndVarId) -> &InductionVar<L>

Retrieves an induction variable given its Id.

pub fn induction_vars<'b>(
&'b self
) -> impl Iterator<Item = (IndVarId, &'b InductionVar<'a, L>)>

Iterates over induction variables.

pub fn set_iteration_dim(&mut self, inst: InstId, dim: DimId) -> bool

Sets a dimension as an iteration dimension for an instruction. Indicates if the iteration dimension was not aleady present in the set.

pub fn add_thread_dim(&mut self, dim: DimId) -> bool

Adds a thread dimension. Indicates if the the dimension was not already present in the set.

pub fn dim_mappings(
&self
) -> impl Iterator<Item = &DimMapping> + '_

Returns the list of dimensions mapping.

pub fn dim_mapping(&self, id: DimMappingId) -> &DimMapping

Retrives a dimension mapping given its ID.

pub fn find_mapping(&self, lhs: DimId, rhs: DimId) -> Option<DimMappingId>

Tries to find a mapping between two dimensions.

impl<'a> Function<'a, ()>

pub fn add_inst(
&mut self,
op: Operator<'a, ()>,
iter_dims: HashSet<DimId>
) -> Result<InstId, Error>

Adds an instruction to the function.

pub fn add_mem_block(&mut self, size: u32) -> MemId

Allocates a new memory block.

pub fn add_logical_dim(
&mut self,
size: Size<'a>,
tiling_factors: VecSet<u32>,
possible_tile_sizes: Vec<VecSet<u32>>
) -> Result<(LogicalDimId, Vec<DimId>), Error>

Create a new logical dimension composed of multiple dimensions to implement strip-mining.

pub fn map_dimensions(&mut self, dims: [DimId; 2]) -> DimMappingId

Specifies two dimensions must have the same size have can be used for point-to-point communication.