Struct telamon::ir::Function

pub struct Function<'a, L = LoweringMap> { /* fields omitted */ }

Describes a function and the set of its possible implementation.

Methods

impl<'a, L> Function<'a, L>

pub fn new(signature: &'a Signature, device: &'a Device) -> Self

Creates a new function.

pub fn signature(&self) -> &'a Signature

Returns the function signature.

pub fn device(&self) -> &'a Device

Returns the device the function is compiled for.

pub fn add_ind_var(&mut self, ind_var: InductionVar<'a, L>) -> IndVarId

Adds an induction variable.

pub fn insts<'b>(
&'b self
) -> impl Iterator<Item = &'b Instruction<'a, L>>

Returns the list of instructions of the function.

pub fn dims<'b>(
&'b self
) -> impl Iterator<Item = &'b Dimension<'a>> + Clone

Returns the list of dimensions of the function.

pub fn static_dims<'b>(
&'b self
) -> impl Iterator<Item = &'b Dimension<'a>>

Returns the list of stastic dimensions in the function.

pub fn thread_dims(
&self
) -> impl Iterator<Item = &Dimension<'a>>

Returns the list of thread dimensions.

pub fn inst(&self, id: InstId) -> &Instruction<'a, L>

Returns an instruction given its id.

pub fn dim(&self, id: DimId) -> &Dimension<'a>

Retuns a dimension given its id.

pub fn mem_blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b Block>

Returns the list of memory blocks. The block with id i is in i-th position.

pub fn mem_insts<'b>(
&'b self
) -> impl Iterator<Item = &'b Instruction<'a, L>> + 'b

Iterates over memory instructions.

pub fn internal_mem_blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b InternalBlock>

Returns the internal memory blocks.

pub fn mem_block(&self, id: MemId) -> &Block

Returns a memory block given its id.

pub fn internal_mem_block(&self, id: InternalId) -> &InternalBlock

Returns an internal memory block given its id.

pub fn induction_var(&self, id: IndVarId) -> &InductionVar<L>

Retrieves an induction variable given its Id.

pub fn induction_vars<'b>(
&'b self
) -> impl Iterator<Item = (IndVarId, &'b InductionVar<'a, L>)>

Iterates over induction variables.

pub fn set_iteration_dim(&mut self, inst: InstId, dim: DimId) -> bool

Sets a dimension as an iteration dimension for an instruction. Indicates if the iteration dimension was not aleady present in the set.

pub fn add_thread_dim(&mut self, dim: DimId) -> bool

Adds a thread dimension. Indicates if the the dimension was not already present in the set.

impl<'a> Function<'a, ()>

pub fn add_inst(
&mut self,
op: Operator<'a, ()>,
iter_dims: HashSet<DimId>
) -> Result<InstId, Error>

Adds an instruction to the function.

pub fn add_dim(&mut self, size: Size<'a>) -> Result<DimId, Error>

Creates a new dimension.

pub fn add_mem_block(&mut self, size: u32) -> InternalId

Allocates a new memory block.

impl<'a> Function<'a>

pub fn block(&self, id: BBId) -> &BasicBlock<'a>

Returns a BasicBlock given its id.

pub fn blocks<'b>(
&'b self
) -> impl Iterator<Item = &'b BasicBlock<'a>>

Lists all BasicBlocks.