Struct telamon::ir::Instruction

pub struct Instruction<'a> { /* fields omitted */ }

Represents an instruction.

Methods

impl<'a> Instruction<'a>

pub fn new(
operator: Operator<'a>,
id: InstId,
iter_dims: HashSet<Id>,
device: &Device
) -> Instruction<'a>

Creates a new instruction and type-check the operands.

pub fn operands(&self) -> Vec<&Operand<'a>>

Returns the list of operands of an Instruction.

pub fn t(&self) -> Type

Returns the type of the value produced by an instruction.

pub fn operator(&self) -> &Operator

Returns the operator of the instruction.

pub fn id(&self) -> InstId

Returns the InstId representing the instruction.

pub fn has_side_effects(&self) -> bool

Returns true if the instruction has side effects.

pub fn lower_layout(
&mut self,
ld_idx: Operand<'a>,
ld_pattern: AccessPattern,
st_idx: Operand<'a>,
st_pattern: AccessPattern
)

Applies the lowering of a layout to the instruction.

pub fn lower_dim_map(
&mut self,
op_id: usize,
new_src: InstId,
new_dim_map: DimMap
)

Lowers the DimMap of an operand into an access to a temporary memory.

pub fn dim_maps_to_lower(&self, lhs: Id, rhs: Id) -> Vec<usize>

Indicates the operands for wich a DimMap must be lowered if lhs and rhs are not mapped.

pub fn as_mem_inst(&self) -> Option<&Instruction>

Returns 'self' if it is a memory instruction.

pub fn as_reduction(&self) -> Option<(InstId, &DimMap, &[Id])>

Indicates if the instruction performs a reduction.

pub fn is_reduction_common_dim(&self, init: InstId, dim: Id) -> bool

Returns 'true' if self is a reduction initialized by init, and if 'dim' should have the same nesting with 'init' that with 'self'.

pub fn merge_dims(&mut self, lhs: Id, rhs: Id)

Rename a dimension to another ID.

pub fn iteration_dims(&self) -> &HashSet<Id>

The list of dimensions the instruction must be nested in.

pub fn add_iteration_dimension(&mut self, dim: Id) -> bool

Adds a new iteration dimension. Indicates if the dimension was not already an iteration dimension.