Struct telamon::search_space::DomainStore

pub struct DomainStore { /* fields omitted */ }

Stores the domains of each variable.

Methods

impl DomainStore

pub fn new(ir_instance: &Function) -> Self

Creates a new domain store and allocates the variables for the given BBMap.

pub fn alloc(&mut self, ir_instance: &Function, new_objs: &NewObjs)

Allocates the choices when new objects are created.

pub fn get_inst_flag(&self, inst: InstId) -> InstFlag

Returns the domain of {name} for the given arguments.

pub fn get_old_inst_flag(&self, inst: InstId, diff: &DomainDiff) -> InstFlag

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_inst_flag(&mut self, inst: InstId, value: InstFlag)

Sets the domain of {name} for the given arguments.

pub fn restrict_inst_flag(
&mut self,
inst: InstId,
value: InstFlag,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_size(&self, dim: DimId) -> NumericSet

Returns the domain of {name} for the given arguments.

pub fn get_old_size(&self, dim: DimId, diff: &DomainDiff) -> NumericSet

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_size(&mut self, dim: DimId, value: NumericSet)

Sets the domain of {name} for the given arguments.

pub fn restrict_size(
&mut self,
dim: DimId,
value: NumericSet,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_thread_dim(&self, dim: DimId) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_is_thread_dim(&self, dim: DimId, diff: &DomainDiff) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_thread_dim(&mut self, dim: DimId, value: Bool)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_thread_dim(
&mut self,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_thread_mapping(&self, lhs: DimId, rhs: DimId) -> ThreadMapping

Returns the domain of {name} for the given arguments.

pub fn get_old_thread_mapping(
&self,
lhs: DimId,
rhs: DimId,
diff: &DomainDiff
) -> ThreadMapping

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_thread_mapping(
&mut self,
lhs: DimId,
rhs: DimId,
value: ThreadMapping
)

Sets the domain of {name} for the given arguments.

pub fn restrict_thread_mapping(
&mut self,
lhs: DimId,
rhs: DimId,
value: ThreadMapping,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_iteration_dim(&self, inst: InstId, dim: DimId) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_is_iteration_dim(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_iteration_dim(&mut self, inst: InstId, dim: DimId, value: Bool)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_iteration_dim(
&mut self,
inst: InstId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_memory_space(&self, var: VarId) -> MemorySpace

Returns the domain of {name} for the given arguments.

pub fn get_old_memory_space(&self, var: VarId, diff: &DomainDiff) -> MemorySpace

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_memory_space(&mut self, var: VarId, value: MemorySpace)

Sets the domain of {name} for the given arguments.

pub fn restrict_memory_space(
&mut self,
var: VarId,
value: MemorySpace,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_mem_space(&self, mem: MemId) -> MemSpace

Returns the domain of {name} for the given arguments.

pub fn get_old_mem_space(&self, mem: MemId, diff: &DomainDiff) -> MemSpace

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_mem_space(&mut self, mem: MemId, value: MemSpace)

Sets the domain of {name} for the given arguments.

pub fn restrict_mem_space(
&mut self,
mem: MemId,
value: MemSpace,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_dim_kind(&self, dim: DimId) -> DimKind

Returns the domain of {name} for the given arguments.

pub fn get_old_dim_kind(&self, dim: DimId, diff: &DomainDiff) -> DimKind

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_dim_kind(&mut self, dim: DimId, value: DimKind)

Sets the domain of {name} for the given arguments.

pub fn restrict_dim_kind(
&mut self,
dim: DimId,
value: DimKind,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_order(&self, lhs: StmtId, rhs: StmtId) -> Order

Returns the domain of {name} for the given arguments.

pub fn get_old_order(
&self,
lhs: StmtId,
rhs: StmtId,
diff: &DomainDiff
) -> Order

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_order(&mut self, lhs: StmtId, rhs: StmtId, value: Order)

Sets the domain of {name} for the given arguments.

pub fn restrict_order(
&mut self,
lhs: StmtId,
rhs: StmtId,
value: Order,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_dim_mapping(&self, lhs: DimId, rhs: DimId) -> DimMapping

Returns the domain of {name} for the given arguments.

pub fn get_old_dim_mapping(
&self,
lhs: DimId,
rhs: DimId,
diff: &DomainDiff
) -> DimMapping

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_dim_mapping(&mut self, lhs: DimId, rhs: DimId, value: DimMapping)

Sets the domain of {name} for the given arguments.

pub fn restrict_dim_mapping(
&mut self,
lhs: DimId,
rhs: DimId,
value: DimMapping,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_tiling_factor(
&self,
logical: LogicalDimId,
dim: DimId
) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_tiling_factor(
&self,
logical: LogicalDimId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_tiling_factor(
&mut self,
logical: LogicalDimId,
dim: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_tiling_factor(
&mut self,
logical: LogicalDimId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_tiling_factor(&self, logical: LogicalDimId) -> Range

Returns the domain of {name} for the given arguments.

pub fn get_old_tiling_factor(
&self,
logical: LogicalDimId,
diff: &DomainDiff
) -> Range

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_tiling_factor(&mut self, logical: LogicalDimId, value: Range)

Sets the domain of {name} for the given arguments.

pub fn restrict_tiling_factor(
&mut self,
logical: LogicalDimId,
value: Range,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_thread_dim_class_counter(&self, dim: DimId) -> Range

Returns the domain of {name} for the given arguments.

pub fn get_old_is_thread_dim_class_counter(
&self,
dim: DimId,
diff: &DomainDiff
) -> Range

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_thread_dim_class_counter(&mut self, dim: DimId, value: Range)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_thread_dim_class_counter(
&mut self,
dim: DimId,
value: Range,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_threads(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_threads(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_threads(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_threads(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_thread_dims(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_thread_dims(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_thread_dims(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_thread_dims(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_iteration_dim_class_counter(
&self,
inst: InstId,
dim: DimId
) -> Range

Returns the domain of {name} for the given arguments.

pub fn get_old_is_iteration_dim_class_counter(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Range

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_iteration_dim_class_counter(
&mut self,
inst: InstId,
dim: DimId,
value: Range
)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_iteration_dim_class_counter(
&mut self,
inst: InstId,
dim: DimId,
value: Range,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_mem_size(&self, mem: MemId, lhs: DimId, rhs: DimId) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_mem_size(
&self,
mem: MemId,
lhs: DimId,
rhs: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_mem_size(
&mut self,
mem: MemId,
lhs: DimId,
rhs: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_mem_size(
&mut self,
mem: MemId,
lhs: DimId,
rhs: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_mem_size(&self, mem: MemId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_mem_size(&self, mem: MemId, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_mem_size(&mut self, mem: MemId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_mem_size(
&mut self,
mem: MemId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_shared_mem_used(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_shared_mem_used(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_shared_mem_used(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_shared_mem_used(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_unroll_factor(&self, inst: InstId, dim: DimId) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_unroll_factor(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_unroll_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_unroll_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_unroll_factor(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_unroll_factor(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_unroll_factor(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_unroll_factor(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_inner_vector_factor(
&self,
inst: InstId,
dim: DimId
) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_inner_vector_factor(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_inner_vector_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_inner_vector_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_inner_vector_factor(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_inner_vector_factor(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_inner_vector_factor(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_inner_vector_factor(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_outer_vector_factor(
&self,
inst: InstId,
dim: DimId
) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_outer_vector_factor(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_outer_vector_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_outer_vector_factor(
&mut self,
inst: InstId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_outer_vector_factor(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_outer_vector_factor(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_outer_vector_factor(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_outer_vector_factor(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_num_block_dims(&self, inst: InstId, dim: DimId) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_num_block_dims(
&self,
inst: InstId,
dim: DimId,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_num_block_dims(
&mut self,
inst: InstId,
dim: DimId,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_num_block_dims(
&mut self,
inst: InstId,
dim: DimId,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_block_dims(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_block_dims(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_block_dims(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_block_dims(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_nested_inst(&self, dim: DimId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_nested_inst(
&self,
dim: DimId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_nested_inst(&mut self, dim: DimId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_nested_inst(
&mut self,
dim: DimId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.