Struct telamon::search_space::DomainStore

pub struct DomainStore { /* fields omitted */ }

Stores the domains of each variable.

Methods

impl DomainStore

pub fn new(ir_instance: &Function) -> Self

Creates a new domain store and allocates the variables for the given BBMap.

pub fn alloc(&mut self, ir_instance: &Function, new_objs: &NewObjs)

Allocates the choices when new objects are created.

pub fn get_is_thread_dim(&self, dim: Id) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_is_thread_dim(&self, dim: Id, diff: &DomainDiff) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_thread_dim(&mut self, dim: Id, value: Bool)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_thread_dim(
&mut self,
dim: Id,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_dim_kind(&self, dim: Id) -> DimKind

Returns the domain of {name} for the given arguments.

pub fn get_old_dim_kind(&self, dim: Id, diff: &DomainDiff) -> DimKind

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_dim_kind(&mut self, dim: Id, value: DimKind)

Sets the domain of {name} for the given arguments.

pub fn restrict_dim_kind(
&mut self,
dim: Id,
value: DimKind,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_order(&self, lhs: BBId, rhs: BBId) -> Order

Returns the domain of {name} for the given arguments.

pub fn get_old_order(&self, lhs: BBId, rhs: BBId, diff: &DomainDiff) -> Order

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_order(&mut self, lhs: BBId, rhs: BBId, value: Order)

Sets the domain of {name} for the given arguments.

pub fn restrict_order(
&mut self,
lhs: BBId,
rhs: BBId,
value: Order,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_unroll_factor(&self, inst: InstId, dim: Id) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_unroll_factor(
&self,
inst: InstId,
dim: Id,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_unroll_factor(
&mut self,
inst: InstId,
dim: Id,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_unroll_factor(
&mut self,
inst: InstId,
dim: Id,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_thread_dims(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_thread_dims(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_thread_dims(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_thread_dims(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_dim_mapping(&self, lhs: Id, rhs: Id) -> DimMapping

Returns the domain of {name} for the given arguments.

pub fn get_old_dim_mapping(
&self,
lhs: Id,
rhs: Id,
diff: &DomainDiff
) -> DimMapping

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_dim_mapping(&mut self, lhs: Id, rhs: Id, value: DimMapping)

Sets the domain of {name} for the given arguments.

pub fn restrict_dim_mapping(
&mut self,
lhs: Id,
rhs: Id,
value: DimMapping,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_mem_size(&self, mem: InternalId, lhs: Id, rhs: Id) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_mem_size(
&self,
mem: InternalId,
lhs: Id,
rhs: Id,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_mem_size(
&mut self,
mem: InternalId,
lhs: Id,
rhs: Id,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_mem_size(
&mut self,
mem: InternalId,
lhs: Id,
rhs: Id,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_mem_size(&self, mem: InternalId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_mem_size(&self, mem: InternalId, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_mem_size(&mut self, mem: InternalId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_mem_size(
&mut self,
mem: InternalId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_shared_mem_used(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_shared_mem_used(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_shared_mem_used(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_shared_mem_used(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_increment_num_block_dims(&self, inst: InstId, dim: Id) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_increment_num_block_dims(
&self,
inst: InstId,
dim: Id,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_increment_num_block_dims(
&mut self,
inst: InstId,
dim: Id,
value: Bool
)

Sets the domain of {name} for the given arguments.

pub fn restrict_increment_num_block_dims(
&mut self,
inst: InstId,
dim: Id,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_nested_inst(&self, dim: Id) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_nested_inst(&self, dim: Id, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_nested_inst(&mut self, dim: Id, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_nested_inst(
&mut self,
dim: Id,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_iteration_dim_class_counter(&self, inst: InstId, dim: Id) -> Range

Returns the domain of {name} for the given arguments.

pub fn get_old_is_iteration_dim_class_counter(
&self,
inst: InstId,
dim: Id,
diff: &DomainDiff
) -> Range

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_iteration_dim_class_counter(
&mut self,
inst: InstId,
dim: Id,
value: Range
)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_iteration_dim_class_counter(
&mut self,
inst: InstId,
dim: Id,
value: Range,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_iteration_dim(&self, inst: InstId, dim: Id) -> Bool

Returns the domain of {name} for the given arguments.

pub fn get_old_is_iteration_dim(
&self,
inst: InstId,
dim: Id,
diff: &DomainDiff
) -> Bool

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_iteration_dim(&mut self, inst: InstId, dim: Id, value: Bool)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_iteration_dim(
&mut self,
inst: InstId,
dim: Id,
value: Bool,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_mem_space(&self, mem: Id) -> MemSpace

Returns the domain of {name} for the given arguments.

pub fn get_old_mem_space(&self, mem: Id, diff: &DomainDiff) -> MemSpace

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_mem_space(&mut self, mem: Id, value: MemSpace)

Sets the domain of {name} for the given arguments.

pub fn restrict_mem_space(
&mut self,
mem: Id,
value: MemSpace,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_is_thread_dim_class_counter(&self, dim: Id) -> Range

Returns the domain of {name} for the given arguments.

pub fn get_old_is_thread_dim_class_counter(
&self,
dim: Id,
diff: &DomainDiff
) -> Range

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_is_thread_dim_class_counter(&mut self, dim: Id, value: Range)

Sets the domain of {name} for the given arguments.

pub fn restrict_is_thread_dim_class_counter(
&mut self,
dim: Id,
value: Range,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_block_dims(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_block_dims(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_block_dims(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_block_dims(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_inst_flag(&self, inst: InstId) -> InstFlag

Returns the domain of {name} for the given arguments.

pub fn get_old_inst_flag(&self, inst: InstId, diff: &DomainDiff) -> InstFlag

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_inst_flag(&mut self, inst: InstId, value: InstFlag)

Sets the domain of {name} for the given arguments.

pub fn restrict_inst_flag(
&mut self,
inst: InstId,
value: InstFlag,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_unroll_factor(&self, inst: InstId) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_unroll_factor(
&self,
inst: InstId,
diff: &DomainDiff
) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_unroll_factor(&mut self, inst: InstId, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_unroll_factor(
&mut self,
inst: InstId,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_num_threads(&self) -> HalfRange

Returns the domain of {name} for the given arguments.

pub fn get_old_num_threads(&self, diff: &DomainDiff) -> HalfRange

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_num_threads(&mut self, value: HalfRange)

Sets the domain of {name} for the given arguments.

pub fn restrict_num_threads(
&mut self,
value: HalfRange,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.

pub fn get_thread_mapping(&self, lhs: Id, rhs: Id) -> ThreadMapping

Returns the domain of {name} for the given arguments.

pub fn get_old_thread_mapping(
&self,
lhs: Id,
rhs: Id,
diff: &DomainDiff
) -> ThreadMapping

Returns the domain of {name} for the given arguments. If the domain has been restricted but the change not yet applied, returns the old value.

pub fn set_thread_mapping(&mut self, lhs: Id, rhs: Id, value: ThreadMapping)

Sets the domain of {name} for the given arguments.

pub fn restrict_thread_mapping(
&mut self,
lhs: Id,
rhs: Id,
value: ThreadMapping,
diff: &mut DomainDiff
) -> Result<(), ()>

Restricts the domain of {name} for the given arguments. Put the old value in diff and indicates if the new domain is failed.