Trait telamon::search_space::Domain

pub trait Domain: Copy + Eq {
    fn is_failed(&self) -> bool;

    fn is_constrained(&self) -> bool;

    fn contains(&self, other: Self) -> bool;

    fn restrict(&mut self, other: Self);

    fn intersects(&self, other: Self) -> bool
    where
        Self: Sized,
    { ... }

    fn is(&self, other: Self) -> Trivalent
    where
        Self: Sized,
    { ... }
}

Required methods

fn is_failed(&self) -> bool

Indicates if the domain is empty.

fn is_constrained(&self) -> bool

Indicates if the domain contains a single alternative.

fn contains(&self, other: Self) -> bool

Indicates if the domain contains another.

fn restrict(&mut self, other: Self)

Restricts the domain to the intersection with other.

Loading content...
Provided methods

fn intersects(&self, other: Self) -> bool where
Self: Sized,

Indicates if the domain has an alternatve in common with other.

fn is(&self, other: Self) -> Trivalent where
Self: Sized,

Indicates if the domain is equal to another domain.

Loading content...